### 2024-10-19 #7374 @editorialbot
[PRE REVIEW]: hetGPy: Heteroskedastic Gaussian Process Modeling in Python
**Submitting author:** @davidogara (<a href="http://orcid.org/0000-0002-1957-400X">David O’Gara</a>)
**Repository:** https://github.com/davidogara/hetGPy
**Branch with paper.md** (empty if default branch):
**Version:** v1.1
