### 2024-10-29 bf94246
thin out requirements
