+++ requirements.txt
@@ -1,6 +1,5 @@
-matplotlib>=3.9.1,<4.0.0
-numba>=0.60.0,<0.61.0
-numpy>=1.20.2,<2.0.0
+matplotlib>=3.6
+numpy>=1.20.2
 scipy>=1.14.0
 tqdm>=4.66.4,<5.0.0
 joblib
