### 2024-11-11 #7374 @matthewfeickert
:wave: @Bobby-Huggins You've been identified as a potential good reviewer for this submission. Would you be willing to review it? For an overview of what reviewing for JOSS is like please see the Reviewing for JOSS documentation: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html Please read this while considering the request to become a reviewer.

If you are interested in reviewing we would like reviewers to aim to have their preliminary reviews (not fully finished!) completed within 3-4 weeks. We understand that everyone is busy and so even if you are interested your schedule might prohibit you from accepting this review at this time. If you have any questions please feel free to ask me. I would be happy to discuss them with you.
