### 2024-11-11 #7374 @matthewfeickert
> I am actually in the process of joining Roman Garnett's lab here at WashU, so this is probably a conflict of interest?

Thanks for checking this in advance @Bobby-Huggins. Yes, if you have already joined the lab and are currently working with Roman then this would hit the [JOSS Conflict of Interest Policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). Though if you haven't yet formally started work and you wouldn't join the lab until you're already well into the review then I think this would be okay.
