### 2024-11-23 #7518 @editorialbot
[REVIEW]: hetGPy: Heteroskedastic Gaussian Process Modeling in Python
**Submitting author:** @davidogara (<a href="http://orcid.org/0000-0002-1957-400X">David O’Gara</a>)
**Repository:** https://github.com/davidogara/hetGPy
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.3
