### 2024-12-03 #7518 @Edenhofer
One the code itself: The software seems well written and has good documentation. The tests and the documentation could be more exposed though (run tests in a CI, host the documentation online). In the following is my review:

* Why does the README explicitly states that only numpy and scipy are dependencies when the requirements.txt lists joblib, matplotlib, and tqdm as additional dependencies? This is still a short list so I don’t think this invalidates the statement that hetGPy has few dependencies but why not state the others too? If they are optional dependencies, they should be marked as such.
* Why is the folder `notebooks` separate from `examples`? At a quick glance both seem to serve the same purpose.
* It would be great if the documentation would be hosted online and linked to in the README.
* It would be great if the test and ideally also the examples would be run automatically in e.g. a GitHub CI.
* Please add or reference community guidelines.
* Could you provide a script to reproduce the performance timings from the paper?
