### 2024-12-03 bf6e26b
fix typos in test suite
