+++ .github/workflows/run-tests.yml
@@ -17,20 +17,15 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: '3.10' 
-      - name: Install cibuildwheel
-        run: python -m pip install cibuildwheel==2.20.0
       - name: Install dependencies
         run: python -m pip install -r requirements.txt
+      - name: Install dev dependencies
+        run: python -m pip install -r requirements_dev.txt
       - name: Handle eigen
         run: git submodule update --init --recursive
-
+      - name: Build package
+        run: pip install -e .
       - name: Build wheels
-        run: python -m cibuildwheel --output-dir wheelhouse
-        # to supply options, put them in 'env', like:
-        # env:
-        #   CIBW_SOME_OPTION: value
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: cibw-wheels-${{ matrix.os }}-${{ strategy.job-index }}
-          path: ./wheelhouse/*.whl
+        run: python -m build
+      - name: Test
+        run: pytest
