### 2024-12-03 be8ffa1
add R installation script for pytest
