### 2024-12-03 e7449d2
add install_R.py to run-tests.yml
