### 2024-12-03 10b6f2b
make R libraries install to personal library
