### 2024-12-03 #7518 @matthewfeickert
@Edenhofer Please open up Issues on https://github.com/davidogara/hetGPy and link them back to the review issue instead of iterating on your review comments here.
