### 2024-12-03 50930a9
add requirements for docs
