+++ docs/requirements_docs.txt
@@ -1,0 +1,6 @@
+dash>=2.14.1
+ipykernel>=6.27.1
+jupyter>=1.0.0
+sphinx>=7.2.6
+numpydoc>=1.6.0
+sphinx-rtd-theme>=2.0.0
