### 2024-12-03 5db5069
add python requirements to readthedocs
