+++ docs/requirements_docs.txt
@@ -4,3 +4,4 @@ jupyter>=1.0.0
 sphinx>=7.2.6
 numpydoc>=1.6.0
 sphinx-rtd-theme>=2.0.0
+nbsphinx
