### 2024-12-03 4ca1862
update readthedocs to py311 and ubuntu 20.04
