+++ .readthedocs.yaml
@@ -7,9 +7,9 @@ version: 2
 
 # Set the OS, Python version and other tools you might need
 build:
-  os: ubuntu-24.04
+  os: "ubuntu-20.04"
   tools:
-    python: "3.13"
+    python: "3.11"
     # You can also specify other tool versions:
     # nodejs: "23"
     # rust: "1.82"
