### 2024-12-03 6fdfd73
update docs requirements
