### 2024-12-03 59fb231
update doc requirements
