### 2024-12-03 91359ad
remove doc requirements
