+++ docs/requirements_docs.txt
@@ -5,10 +5,3 @@ sphinx>=7.2.6
 numpydoc>=1.6.0
 sphinx-rtd-theme>=2.0.0
 nbsphinx>=0.9.5
-
-# add package requirements
-matplotlib>=3.6
-numpy>=1.20.2
-scipy>=1.14.0
-tqdm>=4.66.4,<5.0.0
-joblib
