+++ .readthedocs.yaml
@@ -30,3 +30,4 @@ sphinx:
 python:
   install:
   - requirements: docs/requirements_docs.txt
+  - requirements: requirements.txt
