### 2024-12-03 095be8c
build package during installation
