### 2024-12-03 6688639
build package during installation
