+++ .readthedocs.yaml
@@ -12,6 +12,8 @@ build:
     python: "3.11"
   jobs:
     pre_install:
+      - git submodule update --init --recursive
+      - pip install -r requirements.txt
       - pip install -e .
 
 # Build documentation in the "docs/" directory with Sphinx
