### 2024-12-04 davidogara/hetGPy#8 @Edenhofer
Clarify dependencies
Thanks for developing this software. I have a minor question about the list of dependencies. Why does the README states that only numpy and scipy are dependencies when the requirements.txt lists joblib, matplotlib, and tqdm as additional dependencies? This is still a short list so I don’t think this invalidates the statement that hetGPy has few dependencies but why not state the others too?

Part of https://github.com/openjournals/joss-reviews/issues/7518 .
