### 2024-12-04 davidogara/hetGPy#9 @Edenhofer
`notebooks` vs `examples`
This might be a stupid question but why is the `notebooks` folder distinct from the `examples` folder. Both seem very much related.

Part of https://github.com/openjournals/joss-reviews/issues/7518 .
