### 2024-12-04 davidogara/hetGPy#10 @Edenhofer
Online Documentation
It would be great if the documentation would be hosted online and linked to in the README.

Part of https://github.com/openjournals/joss-reviews/issues/7518 .
