### 2024-12-04 davidogara/hetGPy#11 @Edenhofer
Automate Tests
It would be great if the test and ideally also the examples would be run automatically in e.g. a GitHub CI.

Part of https://github.com/openjournals/joss-reviews/issues/7518 .
