### 2024-12-04 davidogara/hetGPy#12 @Edenhofer
Community Guidelines
Please add or reference community guidelines.

Part of https://github.com/openjournals/joss-reviews/issues/7518 .
