### 2024-12-04 #7518 @davidogara
Thanks for sending these along @Edenhofer! I will respond to each of the issues in the repository, hopefully in the next few days.
