### 2024-12-04 4b524c7
remove R deps from test_allocate_mult and test_auto_bounds
