### 2024-12-04 fd2eba8
add tests for duplicated, EI, and find_reps
