+++ tests/test_find_reps.py
@@ -1,11 +1,10 @@
 from hetgpy.find_reps import find_reps
-from rpy2.robjects import default_converter, numpy2ri
 import numpy as np
-from time import time
 from scipy.io import loadmat
-from rpy2.robjects.packages import importr
-hetGP_R = importr('hetGP')
-
+from tests.utils import read_yaml
+R_compare = read_yaml('tests/R/results/test_find_reps.yaml')
+R_compare['mcycle']['X0'] = np.array(R_compare['mcycle']['X0']).reshape(-1,1)
+R_compare['SIR']['X0'] = np.array(R_compare['SIR']['X0']).reshape(2,-1).T
 
 def test_find_reps_on_mcycle():
 
@@ -19,29 +18,19 @@ def test_find_reps_on_mcycle():
         return_Zlist=True,
         normalize=False
     )
-    
-    # run in R
-    np_cv_rules = default_converter + numpy2ri.converter
-    with np_cv_rules.context():
-        test_R = hetGP_R.find_reps(
-            X = X,
-            Z = Z,
-            rescale=False, 
-            return_Zlist=True,
-            normalize=False
-        )
-        
+    test_R = R_compare['mcycle']
     for key in ('X0', 'Z0', 'mult', 'Z', 'Zlist'):
         if key=="Zlist":
             for k in test[key]:
-                assert np.allclose(test[key][k], test_R[key][str(k+1)])
+                assert np.allclose(test[key][k], np.array(test_R[key][str(k+1)]))
         else:
             assert np.allclose(test[key], test_R[key])
 
 def test_find_reps_SIR():
     SIR = loadmat('tests/data/SIR.mat')
-    X = SIR['X']
-    Z = SIR['Y']
+    
+    X = SIR['X'][0:200,:]
+    Z = SIR['Y'][0:200]
     test = find_reps(
         X = X,
         Z = Z,
@@ -51,20 +40,13 @@ def test_find_reps_SIR():
     )
     
     # run in R
-    np_cv_rules = default_converter + numpy2ri.converter
-    with np_cv_rules.context():
-        test_R = hetGP_R.find_reps(
-            X = X,
-            Z = Z,
-            rescale=False, 
-            return_Zlist=True,
-            normalize=False
-        )
+    
+    test_R = R_compare['SIR']
         
     for key in ('X0', 'Z0', 'mult', 'Z', 'Zlist'):
         if key=="Zlist":
             for k in test[key]:
-                assert np.allclose(test[key][k], test_R[key][str(k+1)])
+                assert np.allclose(test[key][k], np.array(test_R[key][str(k+1)]))
         else:
             assert np.allclose(test[key], test_R[key])
 
