+++ tests/test_duplicated.py
@@ -1,20 +1,20 @@
 from hetgpy.utils import duplicated
 import numpy as np
-from rpy2.robjects import r
-
+from tests.utils import read_yaml
+R_data = read_yaml('tests/R/results/test_duplicated.yaml')
+for key in R_data:
+    R_data[key] = np.array(R_data[key])
 def test_duplicates():
     # from help(duplicated) in R
-    r("x <- c(9:20, 1:5, 3:7, 0:8)")
-    x = np.array(r("x"))
-    test1 = (duplicated(x) == np.array(r("duplicated(x)")).astype(bool)).all()
-    test2 = (duplicated(x,fromLast = True) == np.array(r("duplicated(x,fromLast = T)")).astype(bool)).all()
+    x = R_data['x']
+    test1 = (duplicated(x) == R_data['D1'].astype(bool)).all()
+    test2 = (duplicated(x,fromLast = True) == R_data['D1_L'].astype(bool)).all()
     assert test1
     assert test2
 
     # 2d example
-    r('X <- matrix(c(c(1,2,3),c(1,2,3),c(4,5,6)),nrow=3,byrow=T)')
-    X = np.array(r('X'))
-    test3 = (duplicated(X) == np.array(r('duplicated(X)')).astype(bool)).all()
-    test4 = (duplicated(X,fromLast=True) == np.array(r('duplicated(X,fromLast=T)')).astype(bool)).all()
+    X = R_data['X']
+    test3 = (duplicated(X) == R_data['D2'].astype(bool)).all()
+    test4 = (duplicated(X,fromLast=True) == R_data['D2_l'].astype(bool)).all()
     assert test3
     assert test4
