### 2024-12-04 1cceaa9
add mcycle and SIR tests
