+++ tests/test_homGP_mcycle.py
@@ -1,11 +1,7 @@
-import os
-import pandas as pd
 from scipy.io import loadmat
 import numpy as np
 from hetgpy.homGP import homGP
-import matplotlib as mpl
-import matplotlib.pyplot as plt
-from scipy.stats import norm
+from tests.utils import read_yaml
 
 def test_homGP_mcycle():
     '''
@@ -34,100 +30,20 @@ def test_homGP_mcycle():
     )
 
     # Compare to R:
-    '''
-    # Test #1 -- mcycle Gaussian
-    hom <- mleHomGP(mcycle$times, mcycle$accel, covtype = "Gaussian",
-                    maxit = 100, settings = list('factr'=10))
-
-    Xgrid <- matrix(seq(0, 60, length = 301), ncol = 1)
-    p <- predict(x = Xgrid, object = hom)
-
-    ## motopred
-    R.matlab::writeMat(
-    # fit
-    theta = hom$theta, g = hom$g,
-    beta0 = hom$beta0, Ki = hom$Ki,
-    X0 = hom$X0, Z0 = hom$Z0,covtype = hom$covtype,
-    nu_hat = hom$nu_hat, trendtype = hom$trendtype,
-    # preds
-    mean = p$mean,
-    sd2 = p$sd2,
-    nugs = p$nugs,
-    # filepath
-    con = 'data/homGP_mcycle.mat')
-    '''
-    mat = loadmat('tests/data/homGP_mcycle.mat')
-    preds_mean_R = mat['mean']
-    preds_sd_R   = mat['sd2']
+    R_compare    = read_yaml('tests/R/results/test_homGP_mcycle.yaml')
+    preds_mean_R = R_compare['mean']
+    preds_sd_R   = R_compare['sd2']
     
     
     # model pars
-    assert np.allclose(model['theta'],mat['theta'])
-    assert np.allclose(model['g'],mat['g'])
-    assert np.allclose(model['Ki'], mat['Ki'],atol=1e-7)
+    assert np.allclose(model['theta'],R_compare['theta'])
+    assert np.allclose(model['g'],R_compare['g'])
     
     # predictions
     assert np.allclose(preds['mean'],preds_mean_R,atol=1e-6)
     assert np.allclose(preds['sd2'],preds_sd_R,atol=1e-6)
-    #make_plot(preds,mat,xgrid,X,Z,save_plot=True)
+    
     return
-
-     
-
-
-def make_plot(preds,mat,xgrid,X,Z, save_plot = True):
-        '''
-        Make a plot comparing the hetGP to hetGPy.
-        Note that this function assumes a LaTeX installation (due to usetex = True)
-        '''
-        if os.path.exists('mplstyle/latex.mplstyle'): plt.style.use('mplstyle/latex.mplstyle')
-        preds_mean_R = mat['mean']
-        preds_sd_R   = mat['sd2']
-        preds_sd_R_05 = mat['CI_lower']
-        preds_sd_R_95 = mat['CI_upper']
-        df = pd.DataFrame(
-            {
-            # python results
-            'hetGPy': preds['mean'],
-            'hetGPy-0.05': norm.ppf(0.05, loc = preds['mean'], scale = np.sqrt(preds['sd2'] + preds['nugs'])).squeeze(),
-            'hetGPy-0.95': norm.ppf(0.95, loc = preds['mean'], scale = np.sqrt(preds['sd2'] + preds['nugs'])).squeeze(),
-            # R results
-            'hetGP': preds_mean_R,
-            'hetGP-0.05': preds_sd_R_05,
-            'hetGP-0.95': preds_sd_R_95
-            },
-            index=xgrid.squeeze()
-        )
[…]
