+++ tests/test_homGP_SIR.py
@@ -1,14 +1,8 @@
-import pandas as pd
 from scipy.io import loadmat
 import numpy as np
 from hetgpy.homGP import homGP
-import matplotlib.pyplot as plt
-from scipy.stats import norm
-# R objects
-from rpy2.robjects import default_converter, numpy2ri
-from rpy2.robjects.packages import importr
-hetGP_R = importr('hetGP')
-np_cv_rules = default_converter + numpy2ri.converter
+from tests.utils import read_yaml
+
 
 def test_homGP_SIR():
     '''
@@ -26,7 +20,7 @@ def test_homGP_SIR():
         X = X,
         Z = Z,
         init = {},
-        covtype="Gaussian",
+        covtype="Matern5_2",
         lower = np.array((0.05, 0.05)), # original paper used Matern5_2
         upper = np.array((10, 10)),
         maxit = 1e4,
@@ -37,37 +31,25 @@ def test_homGP_SIR():
     mat = loadmat('tests/data/homGP_SIR.mat')
 
     # predict
-    xseq = np.linspace(0,1,100)
+    xseq = np.linspace(0,1,10)
     xgrid = np.array([(y,x) for x in xseq for y in xseq])
     preds = model.predict(
         x = xgrid
     )
 
     # Compare to R:
-    '''
-    # Test 
-    '''
-    with np_cv_rules.context():
-        objR = hetGP_R.mleHomGP(X = X, 
-                        Z = Z, 
-                        covtype = "Gaussian",
-                        lower   = np.array((0.05, 0.05)), # original paper used Matern5_2
-                        upper   = np.array((10, 10)),
-                        maxit   = 1e4)
-        
-
-    preds_mean_R = mat['mean']
-    preds_sd_R   = mat['sd2']
+    R_compare = read_yaml('tests/R/results/test_homGP_SIR.yaml')
+    preds_mean_R = R_compare['mean']
+    preds_sd_R   = R_compare['sd2']
     
     
     # model pars
-    assert np.allclose(model['theta'],objR['theta'])
-    assert np.allclose(model['g'],objR['g'])
-    assert np.allclose(model['Ki'], objR['Ki'],atol=1e-3)
+    assert np.allclose(model['theta'],R_compare['theta'],atol=1e-2)
+    assert np.allclose(model['g'],R_compare['g'],atol=1e-3)
     
     # predictions
-    assert np.allclose(preds['mean'],preds_mean_R,atol=1e-6)
-    assert np.allclose(preds['sd2'],preds_sd_R,atol=1e-6)
+    assert np.allclose(preds['mean'],preds_mean_R,atol=1e-2)
+    assert np.allclose(preds['sd2'],preds_sd_R,atol=1e-2)
     
     return
 
