### 2024-12-04 9fea191
add README and utils for testing dir
