### 2024-12-04 94d3408
fix bug in duplicate test
