+++ tests/R/results/test_horizon.yaml
@@ -1,0 +1,193 @@
+theta: 0.2571488
+g: 3.2621795e-06
+Delta:
+- -7.4285792
+- -7.5389267
+- -7.8281253
+- -8.0252698
+- -8.2418305
+- -9.1394078
+- -9.1799094
+- -9.16799
+- -9.0341927
+- -8.8796893
+- -8.5096145
+- -7.8669666
+- -7.5332569
+- -7.3623864
+- -7.0569486
+- -6.7596981
+- -6.4988882
+- -5.0346806
+- -4.5696189
+- -4.3297425
+- -3.4310847
+- -3.2041636
+- -2.6752079
+- -2.5344611
+- -2.3916366
+- -2.2747814
+- -2.1107089
+- -2.0513873
+- -1.9686025
+- -1.8969319
+- -1.7289849
+- -1.7104304
+- -1.760238
+- -1.9267073
+- -1.9780822
+- -2.0613212
+- -2.2789756
+- -2.354655
+- -2.5213753
+- -2.5476717
+- -2.5793155
+- -2.5790083
+- -2.4069122
+- -2.3593156
+- -2.2209703
+- -2.2019973
+- -2.0816819
+- -2.0164053
+- -1.9294584
+- -1.9011827
+- -1.8629409
+- -1.8445405
+- -1.8557914
+- -1.8825805
+- -1.8495137
+- -1.8697835
+- -1.8404417
+- -1.8095431
+- -1.7981309
+- -1.7583602
+- -1.7502941
+- -1.6679356
+- -1.6518444
+- -1.6033351
+- -1.4778513
+- -1.3590721
+- -1.2715582
+- -1.1373907
+- -1.0640552
+- -1.0075718
+- -1.0228486
+- -1.0235551
+- -1.1016057
+- -1.7733079
+- -2.265328
+- -2.3163331
+- -2.4519036
+- -2.5363166
+- -2.7282535
+- -2.9571905
+- -3.0694527
+- -3.1257872
+- -3.373788
+- -3.444309
+- -3.4891084
+- -3.5385219
+- -3.5791419
+- -3.6193122
+- -3.4770653
+- -3.3882589
+- -3.4990339
+- -3.9178076
+- -4.013868
+- -4.3328021
+Lambda:
+- 0.0005908
+- 0.0005362
+- 0.0003988
+- 0.0003252
+- 0.0002637
+- 0.0001082
+- 0.0001035
+- 0.0001028
+- 0.0001193
+- 0.0001402
+- 0.0002017
+- 0.0003824
+- 0.0005359
+- 0.0006317
+- 0.0008632
+- 0.0011513
+- 0.0015089
+- 0.0066059
+- 0.0103428
+- 0.0130569
+- 0.0324121
+- 0.0398624
+- 0.0686339
+- 0.0801042
+- 0.0922883
+- 0.1049052
+- 0.1175029
+- 0.1295542
+- 0.1406855
+- 0.1507094
+- 0.1776872
+- 0.1804617
+- 0.1715111
+- 0.1469556
+- 0.1373107
+- 0.1276828
+- 0.1024251
+- 0.09592
+- 0.079823
+- 0.0778749
+- 0.0761019
+- 0.076265
+- 0.090067
+- 0.0939227
+- 0.1070905
+- 0.112035
+- 0.1227937
+- 0.1341315
+- 0.1447829
+- 0.1492459
+- 0.1553461
+- 0.1568198
+- 0.1573592
+- 0.156176
+- 0.1558011
+- 0.1561743
+- 0.1605847
+- 0.1627682
+- 0.1649699
+- 0.1713069
+- 0.1764725
[…]
