+++ tests/R/results/test_homGP_SIR.yaml
@@ -1,0 +1,206 @@
+theta:
+- 2.6909348
+- 1.073803
+g: 0.0104147
+mean:
+- 0.0036496
+- 0.0012264
+- -0.0009705
+- -0.0025567
+- -0.0026601
+- -0.0007126
+- 0.0029677
+- 0.0082142
+- 0.0158657
+- 0.0262943
+- 0.0422731
+- 0.0476079
+- 0.0539063
+- 0.0615875
+- 0.0715485
+- 0.0843182
+- 0.0994579
+- 0.116667
+- 0.1366095
+- 0.1593537
+- 0.0778876
+- 0.0894456
+- 0.1026335
+- 0.1179128
+- 0.1362114
+- 0.1580363
+- 0.182872
+- 0.2103415
+- 0.2409699
+- 0.2745396
+- 0.1142308
+- 0.1299536
+- 0.1478072
+- 0.1682938
+- 0.1923696
+- 0.2205381
+- 0.2522611
+- 0.2871639
+- 0.325695
+- 0.3674002
+- 0.1460143
+- 0.1645173
+- 0.18551
+- 0.2095168
+- 0.237501
+- 0.2699613
+- 0.3063742
+- 0.3464299
+- 0.3905547
+- 0.4381084
+- 0.1786093
+- 0.1999209
+- 0.2240127
+- 0.251401
+- 0.2830174
+- 0.3193286
+- 0.3598275
+- 0.404289
+- 0.4531346
+- 0.5055649
+- 0.209668
+- 0.2341776
+- 0.261759
+- 0.2929
+- 0.3284678
+- 0.3688663
+- 0.4135859
+- 0.4624714
+- 0.515924
+- 0.5729947
+- 0.2408818
+- 0.2681364
+- 0.2987228
+- 0.3330985
+- 0.3720555
+- 0.4159224
+- 0.464177
+- 0.5167166
+- 0.573909
+- 0.6346744
+- 0.2681526
+- 0.2969092
+- 0.3291799
+- 0.3653986
+- 0.4062855
+- 0.4520979
+- 0.502306
+- 0.5568523
+- 0.6160721
+- 0.6787909
+- 0.3056142
+- 0.3348667
+- 0.3676844
+- 0.4044795
+- 0.445903
+- 0.4921484
+- 0.542686
+- 0.5974995
+- 0.6568963
+- 0.7196511
+sd2:
+- 5.7984651e-05
+- 2.8210942e-05
+- 2.1287953e-05
+- 1.9120172e-05
+- 1.7772571e-05
+- 1.7882605e-05
+- 1.9969119e-05
+- 2.2754724e-05
+- 2.5464029e-05
+- 3.9055081e-05
+- 1.9176477e-05
+- 6.2985036e-06
+- 5.4768895e-06
+- 5.3840926e-06
+- 5.1431438e-06
+- 5.2352179e-06
+- 5.6230434e-06
+- 5.5055905e-06
+- 5.7441146e-06
+- 1.7527424e-05
+- 1.0547631e-05
+- 5.03215e-06
+- 5.8782515e-06
+- 5.4736685e-06
+- 4.7973112e-06
+- 4.6788088e-06
+- 4.9875626e-06
+- 5.3059925e-06
+- 6.9187081e-06
+- 1.999645e-05
+- 1.097493e-05
+- 5.454969e-06
+- 5.6523896e-06
+- 4.8038887e-06
+- 4.2307793e-06
+- 4.4733885e-06
+- 4.9133833e-06
+- 5.2824803e-06
+- 6.8519779e-06
+- 1.8405675e-05
+- 1.6062052e-05
+- 5.9492655e-06
[…]
