+++ tests/R/results/test_auto_bounds.yaml
@@ -1,0 +1,21 @@
+Gaussian:
+  lower:
+  - 25.7523635
+  - 12.7830689
+  upper:
+  - 4184.664797
+  - 2077.2018956
+Matern3_2:
+  lower:
+  - 3.4289303
+  - 2.4158373
+  upper:
+  - 61.1907633
+  - 43.1116742
+Matern5_2:
+  lower:
+  - 3.4785333
+  - 2.4507848
+  upper:
+  - 54.4242938
+  - 38.3443888
