+++ tests/R/results/test_duplicated.yaml
@@ -1,0 +1,114 @@
+x:
+- 9
+- 10
+- 11
+- 12
+- 13
+- 14
+- 15
+- 16
+- 17
+- 18
+- 19
+- 20
+- 1
+- 2
+- 3
+- 4
+- 5
+- 3
+- 4
+- 5
+- 6
+- 7
+- 0
+- 1
+- 2
+- 3
+- 4
+- 5
+- 6
+- 7
+- 8
+X:
+- 1.0
+- 1.0
+- 4.0
+- 2.0
+- 2.0
+- 5.0
+- 3.0
+- 3.0
+- 6.0
+D1:
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- yes
+- yes
+- yes
+- no
+- no
+- no
+- yes
+- yes
+- yes
+- yes
+- yes
+- yes
+- yes
+- no
+D1_L:
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- yes
+- yes
+- yes
+- yes
+- yes
+- yes
+- yes
+- yes
+- yes
+- yes
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+- no
+D2:
+- no
+- yes
+- no
+D2_L:
+- yes
+- no
+- no
