+++ tests/R/results/test_EI.yaml
@@ -1,0 +1,1 @@
+proposal: 21.2

[…]
