### 2024-12-04 4531446
remove redundant test
