### 2024-12-04 f146672
add yaml to R deps
