### 2024-12-04 99b6af3
Create run-main-tests-on-push.yml
