### 2024-12-04 950357c
Update run-main-tests-on-push.yml
