+++ .github/workflows/run-tests.yml
@@ -1,6 +1,10 @@
-name: Build and Test
+name: Run All Tests
 
 on: workflow_dispatch
+
+env:
+  OPENBLAS_NUM_THREADS: 1
+  OMP_NUM_THREADS: 1
 
 jobs:
   build_wheels:
