+++ .github/workflows/gh-pages.yaml
@@ -5,6 +5,29 @@ on:
     branches: [main] # branch to trigger deployment
 
 jobs:
+  build_wheels:
+    name: Build wheels on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ['ubuntu-latest'] # [ubuntu-latest, windows-latest, macos-13, macos-14]
+
+    steps:
+      - uses: actions/checkout@v4
+      # Used to host cibuildwheel
+      - uses: actions/setup-python@v5
+        with:
+          python-version: '3.10' 
+      - name: Install dependencies
+        run: python -m pip install -r requirements.txt
+      - name: Install dev dependencies
+        run: python -m pip install -r requirements_dev.txt
+      - name: Handle eigen
+        run: git submodule update --init --recursive
+      - name: Build package
+        run: pip install -e .
+  
   pages:
     runs-on: ubuntu-20.04
     environment:
