+++ .github/workflows/gh-pages.yaml
@@ -5,15 +5,16 @@ on:
     branches: [main] # branch to trigger deployment
 
 jobs:
-  build_wheels:
-    name: Build wheels on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      fail-fast: false
-      matrix:
-        os: ['ubuntu-latest'] # [ubuntu-latest, windows-latest, macos-13, macos-14]
-
+  pages:
+    runs-on: ubuntu-20.04
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
+    permissions:
+      pages: write
+      id-token: write
     steps:
+      - uses: sphinx-notes/pages@v3
       - uses: actions/checkout@v4
       # Used to host cibuildwheel
       - uses: actions/setup-python@v5
@@ -27,15 +28,3 @@ jobs:
         run: git submodule update --init --recursive
       - name: Build package
         run: pip install -e .
-  
-  pages:
-    runs-on: ubuntu-20.04
-    environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
-    permissions:
-      pages: write
-      id-token: write
-    steps:
-    - id: deployment
-      uses: sphinx-notes/pages@v3
