+++ .github/workflows/gh-pages.yaml
@@ -14,7 +14,6 @@ jobs:
       pages: write
       id-token: write
     steps:
-      - uses: sphinx-notes/pages@v3
       - uses: actions/checkout@v4
       # Used to host cibuildwheel
       - uses: actions/setup-python@v5
@@ -28,3 +27,4 @@ jobs:
         run: git submodule update --init --recursive
       - name: Build package
         run: pip install -e .
+      - uses: sphinx-notes/pages@v3
