### 2024-12-04 e722fd6
add nbsphinx to dev requirements
