+++ .github/workflows/gh-pages.yaml
@@ -27,4 +27,14 @@ jobs:
         run: git submodule update --init --recursive
       - name: Build package
         run: pip install -e .
-      - uses: sphinx-notes/pages@v3
+      - name: Sphinx build
+        run: |
+          sphinx-build docs docs/build
+      - name: Deploy to GitHub Pages
+        uses: peaceiris/actions-gh-pages@v3
+        if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
+        with:
+          publish_branch: gh-pages
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: docs/build/
+          force_orphan: true
