+++ .github/workflows/gh-pages.yaml
@@ -3,7 +3,8 @@ name: Deploy Sphinx documentation to Pages
 on:
   push:
     branches: [main] # branch to trigger deployment
-
+permissions:
+  contents: write
 jobs:
   pages:
     runs-on: ubuntu-20.04
