+++ .github/workflows/gh-pages.yaml
@@ -3,8 +3,7 @@ name: Deploy Sphinx documentation to Pages
 on:
   push:
     branches: [main] # branch to trigger deployment
-permissions:
-  contents: write
+
 jobs:
   pages:
     runs-on: ubuntu-20.04
@@ -14,6 +13,7 @@ jobs:
     permissions:
       pages: write
       id-token: write
+      contents: write
     steps:
       - uses: actions/checkout@v4
       # Used to host cibuildwheel
