+++ .github/workflows/gh-pages.yaml
@@ -19,7 +19,7 @@ jobs:
       # Used to host cibuildwheel
       - uses: actions/setup-python@v5
         with:
-          python-version: '3.10' 
+          python-version: '3.11' 
       - name: Install dependencies
         run: python -m pip install -r requirements.txt
       - name: Install dev dependencies
