+++ _sources/api/plot.rst.txt
@@ -1,0 +1,8 @@
+====
+plot
+====
+Methods for the `plot` (leave-one-out) module in `hetgpy`
+
+.. automodule:: hetgpy.plot
+   :no-index:
+   :members:
