+++ _sources/api/IMSE.rst.txt
@@ -1,0 +1,6 @@
+====
+IMSE
+====
+
+.. automodule:: hetgpy.IMSE
+   :members:
