+++ _sources/index.rst.txt
@@ -1,0 +1,37 @@
+.. hetGPy documentation master file, created by
+   sphinx-quickstart on Wed Jan 31 10:46:22 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to hetGPy's documentation!
+==================================
+
+.. grid:: 2
+    :gutter: 4
+    .. grid-item-card:: API reference
+        :link: api/index
+        :link-type: doc
+
+        All the gory detail of every function and class.
+
+    .. grid-item-card:: What's new
+        :link: whatsnew
+        :link-type: doc
+
+        See what's hot off the press.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+   :hidden:
+
+   overview
+   api/index
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
