+++ _sources/api/find_reps.rst.txt
@@ -1,0 +1,8 @@
+====================
+find_reps
+====================
+Methods for the `find_reps` module in `hetgpy`
+
+.. automodule:: hetgpy.find_reps
+   :no-index:
+   :members:
