+++ _sources/api/LOO.rst.txt
@@ -1,0 +1,8 @@
+====
+LOO
+====
+Methods for the `LOO` (leave-one-out) module in `hetgpy`
+
+.. automodule:: hetgpy.LOO
+   :no-index:
+   :members:
