+++ _sources/api/hetGP.rst.txt
@@ -1,0 +1,8 @@
+=====
+hetGP
+=====
+
+.. currentmodule:: hetgpy
+
+.. autoclass:: hetgpy.hetGP.hetGP
+   :members:
