+++ _sources/api/index.rst.txt
@@ -1,0 +1,20 @@
+===
+API
+===
+    Welcome to the ``hetGPy`` API
+.. currentmodule:: hetgpy
+
+.. toctree::
+   :hidden:
+
+   hetGP
+   homGP
+   contour
+   covariance_functions
+   optim
+   IMSE
+   find_reps
+   LOO
+   plot
+
+   
