+++ _sources/reference.rst.txt
@@ -1,0 +1,20 @@
+``hetGPy`` API reference
+=================================
+
+.. automodule:: hetgpy.find_reps
+   :members:
+
+.. autoclass:: hetgpy.hetGP.hetGP
+   :members:
+
+.. autoclass:: hetgpy.homGP.homGP
+   :members:
+
+.. automodule:: hetgpy.optim
+   :members:
+
+.. automodule:: hetgpy.IMSE
+   :members:
+
+.. automodule:: hetgpy.LOO
+   :members:
