+++ _sources/api/optim.rst.txt
@@ -1,0 +1,8 @@
+=====
+optim
+=====
+Methods for the `optim` module in `hetgpy`
+
+.. automodule:: hetgpy.optim
+   :no-index:
+   :members:
