+++ .github/workflows/gh-pages.yaml
@@ -1,41 +1,0 @@
-name: Deploy Sphinx documentation to Pages
-
-on:
-  push:
-    branches: [main] # branch to trigger deployment
-
-jobs:
-  pages:
-    runs-on: ubuntu-20.04
-    environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
-    permissions:
-      pages: write
-      id-token: write
-      contents: write
-    steps:
-      - uses: actions/checkout@v4
-      # Used to host cibuildwheel
-      - uses: actions/setup-python@v5
-        with:
-          python-version: '3.11' 
-      - name: Install dependencies
-        run: python -m pip install -r requirements.txt
-      - name: Install dev dependencies
-        run: python -m pip install -r requirements_dev.txt
-      - name: Handle eigen
-        run: git submodule update --init --recursive
-      - name: Build package
-        run: pip install -e .
-      - name: Sphinx build
-        run: |
-          sphinx-build docs docs/build
-      - name: Deploy to GitHub Pages
-        uses: peaceiris/actions-gh-pages@v3
-        if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
-        with:
-          publish_branch: gh-pages
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: docs/build/
-          force_orphan: true
