### 2024-12-05 4aba2b8
add examples to docs
