+++ docs/index.rst
@@ -21,6 +21,13 @@ Welcome to hetGPy's documentation!
         See what's hot off the press.
 
 .. toctree::
+   :glob:
+   :maxdepth: 2
+   :caption: Examples:
+
+   examples/**/index
+   
+.. toctree::
    :maxdepth: 2
    :caption: Contents:
    :hidden:
