### 2024-12-05 25e930b
add index to example folder
