+++ examples/index.rst
@@ -1,0 +1,37 @@
+hetGPy Example Notebooks
+========================
+
+=
+
+.. toctree::
+   :maxdepth: 1
+   :hidden:
+
+   00-hetGPy-Intro.ipynb
+   01-mcycle-example.ipynb
+   02-SIR-example.ipynb
+   03-EI-Example.ipynb
+   04-IMSPE.ipynb
+   05-Update.ipynb
+   06-Sequential-Design.ipynb
+
+.. _hetGPy Intro:
+  ./00-hetGPy-Intro.ipynb
+
+.. _Motorcycle Example:
+  ./01-mcycle-example.ipynb
+
+.. _SIR Model:
+  ./02-SIR-example.ipynb
+
+.. _Expected Improvement:
+  ./03-EI-Example.ipynb
+
+.. _Integrated Mean-Squared Prediction Error:
+  ./04-IMSPE.ipynb
+
+.. _Updating a Model:
+  ./05-Update.ipynb
+
+.. _Sequential Design:
+  ./06-Sequential-Design.ipynb
