+++ examples/index.rst
@@ -1,8 +1,6 @@
 hetGPy Example Notebooks
 ========================
 
-=
-
 .. toctree::
    :maxdepth: 1
    :hidden:
