### 2024-12-05 5c18b82
run sample notebooks on push
