### 2024-12-05 cc15db4
add script to run all notebook examples
