### 2024-12-05 9d33d60
add documentation to README
