### 2024-12-05 davidogara/hetGPy#9 @davidogara
Thanks for pointing this out -- the `notebooks` directory was an old version of what became the `examples` directory, so I have removed it.
