### 2024-12-05 7e05fec
clarify dependencies in README
