### 2024-12-05 davidogara/hetGPy#8 @davidogara
Thanks for pointing this out-- I added a short description to the README stating each dependency and why we included it.
