### 2024-12-05 davidogara/hetGPy#10 @Edenhofer
Awesome! Thank you for fixing this so quickly.

While scrolling through the documentation, I noticed that the lower entries at https://hetgpy.readthedocs.io/en/latest/reference.html#hetgpy.hetGP.hetGP.mleHetGP have some malformed strings in them. Your changes fully addressed my issue but it might be worth cleaning up those formatting issues too while at it.
