### 2024-12-05 davidogara/hetGPy#12 @davidogara
Thanks -- I added `CONTRIBUTING.md` to the repository which encompasses the guidelines for contributing to this project.
