### 2024-12-05 7176dba
update MLE docstrings
