+++ hetgpy/hetGP.py
@@ -527,128 +527,126 @@ class hetGP:
             In the multivariate case, it is possible to give vectors for bounds (resp. scalars) for anisotropy (resp. isotropy)
         noiseControl : dict
             dict with elements related to optimization of the noise process parameters:
-            - ``g_min``, ``g_max`` minimal and maximal noise to signal ratio (of the mean process)
-            - ``lowerDelta``, ``upperDelta`` optional vectors (or scalars) of bounds on ``Delta``, of length ``len(X0)`` (default to ``np.repeat(eps, X0.shape[0])`` and ``np.repeat(noiseControl["g_max"], X0.shape[0])`` resp., or their ``log``) 
-            - ``lowerpX``, ``upperpX`` optional vectors of bounds of the input domain if `pX` is used.
-            - ``lowerTheta_g``, ``upperTheta_g`` optional vectors of bounds for the lengthscales of the noise process if ``linkThetas == 'none'``. Same as for ``theta`` if not provided.
-            - ``k_theta_g_bounds`` if ``linkThetas == 'joint'``, vector with minimal and maximal values for ``k_theta_g`` (default to ``(1, 100)``). See Details.
-            - ``g_bounds`` vector for minimal and maximal noise to signal ratios for the noise of the noise process, i.e., the smoothing parameter for the noise process. (default to ``(1e-6, 1)``).
-    settings : dict 
-            dict for options about the general modeling procedure, with elements:
-            - ``linkThetas`` defines the relation between lengthscales of the mean and noise processes. Either ``'none'``, ``'joint'``(default) or ``'constr'``, see Details.
-            - ``logN``, when ``True`` (default), the log-noise process is modeled.
-            - ``initStrategy`` one of ``'simple'``, ``'residuals'`` (default) and ``'smoothed'`` to obtain starting values for ``Delta``, see Details
-            - ``penalty``  when ``True``, the penalized version of the likelihood is used (i.e., the sum of the log-likelihoods of the mean and variance processes, see References).
-            - ``hardpenalty`` is ``True``, the log-likelihood from the noise GP is taken into account only if negative (default if ``maxit > 1000``).
-            - ``checkHom`` when ``True``, if the log-likelihood with a homoskedastic model is better, then return it.
-            - ``trace`` optional scalar (default to ``0``). If positive, tracing information on the fitting process.
-                If ``1``, information is given about the result of the heterogeneous model optimization.
-                Level ``2`` gives more details. Level ``3`` additionaly displays all details about initialization of hyperparameters.
-            - ``return_matrices`` boolean to include the inverse covariance matrix in the object for further use (e.g., prediction).
-            - ``return_hom`` boolean to include homoskedastic GP models used for initialization (i.e., ``modHom`` and ``modNugs``).
-            - ``factr`` (default to 1e9) and ``pgtol`` are available to be passed to `options` for L-BFGS-B in :func: ``scipy.optimize.minimize``.   
-    eps : float
-        jitter used in the inversion of the covariance matrix for numerical stability
-    init,known :  dict
-        optional lists of starting values for mle optimization or that should not be optimized over, respectively.
-        Values in ``known`` are not modified, while it can happen to these of ``init``, see Details. 
-        One can set one or several of the following:
-        - ``theta`` lengthscale parameter(s) for the mean process either one value (isotropic) or a vector (anistropic)
-        - ``Delta`` vector of nuggets corresponding to each design in ``X0``, that are smoothed to give ``Lambda`` (as the global covariance matrix depends on ``Delta`` and ``nu_hat``, it is recommended to also pass values for ``theta``)
[…]
