### 2024-12-05 davidogara/hetGPy#10 @davidogara
Thanks for pointing this out-- I took a pass at fixing the malformed strings, and the updated documentation is available at: https://hetgpy.readthedocs.io/en/latest/reference.html#hetgpy.hetGP.hetGP.mleHetGP
