### 2024-12-05 davidogara/hetGPy#13 @davidogara
Analysis Script for the paper
From @Edenhofer, part of https://github.com/openjournals/joss-reviews/issues/7518

Could you provide a script to reproduce the performance timings from the paper?
