### 2024-12-05 davidogara/hetGPy#13 @davidogara
Our analysis scripts to generate the results in the paper is found at:

https://github.com/davidogara/hetGPy/tree/main/paper/analysis
