### 2024-12-05 davidogara/hetGPy#14 @davidogara
Paper Comments
From @Edenhofer, available at https://github.com/openjournals/joss-reviews/issues/7518#issuecomment-2515170196:

The paper is well written and contains all the essential information that a reader might want to know about the software. I have no major points of feedback regarding the scientific content of the paper. However, I would suggest to run a grammar checker on the text. There are several duplications of partial sentences in the paper. In the following are a few minor comments about a few specific sections.

- l21 heteroskedasticity is defined more generally as non equal variance noise (not necessarily input dependent)
- l22 “these both of these”; language
- l21ff IMO the sentence is a bit overloaded and hard to read. Please rephrase.
- l23 here and in the following, please be more specific with what you mean with “replicas”/“replication”. I think you mean multiple observations at the same location. I would suggest spelling this out instead of using the term “replication”.
- l46 the default behavior of a library is not too relevant. Rather say whether these libraries can do that. It’s been a while since I’ve - used them but AFAIK some of these libraries support learning the mean and variance.
- l47 “as in common in”; language
- l49 Could the authors briefly elaborate on why hetGPy is more computationally efficient.
- l55 “several orders of magnitude” the table shows a significant improvement but not over “several orders of magnitude”
- l66 the second citation should probably not be in brackets
- l72 “see that the for the SK case”; language
