### 2024-12-05 davidogara/hetGPy#14 @davidogara
Thanks for pointing these out and for the close read @Edenhofer -- I opened this issue so we have a place to discuss (and I hope to address these soon)
