### 2024-12-07 davidogara/hetGPy#15 @DanWaxman
Analysis Results
Thanks for including the analysis scripts in #13. I ran the Python scripts locally and obtained similar results.

I have a few comments about the table formatting, but I figured if I enumerate them I should just make a pull request with it. I will make that shortly -- feel free to accept/reject these changes, of course.

Here are some other thoughts about the testing scripts and the results in the paper:
- There are some unused variables. For example, https://github.com/davidogara/hetGPy/blob/7176dba4cc035baf065efb8e7ce1d00258eac0c6/paper/analysis/py/run_GPyTorch_with_botorch_SK.py#L65
and
https://github.com/davidogara/hetGPy/blob/7176dba4cc035baf065efb8e7ce1d00258eac0c6/paper/analysis/py/run_GPyTorch_with_botorch_SK.py#L75
are both unused.
- No computational resources are mentioned. I understand that the "absolute" numbers are not the important part, but this still seems like good practice when reporting wall-clock times. For example, on my computer, all methods run in ~2/3 of the time. This also need not happen uniformly: for example, hetGPy runs faster than gPyTorch (SK) on Sphere-6D with the Matern kernel for me.
- Similarly, I doubt that there is much variance in the wall-clock time of these scripts, but it still seems best practice to run each script several times and report, e.g., the median.

Part of https://github.com/openjournals/joss-reviews/issues/7518
