### 2024-12-07 davidogara/hetGPy#14 @DanWaxman
Also as part of https://github.com/openjournals/joss-reviews/issues/7518, I will add my paper comments here.

- There are unused affiliations -- looking briefly, it looks like these should be attached to Ross A Hammond.
- (L17) The citation here should be at the end of the sentence.
- (L21-L25) I also think this sentence should be broken up into several sentences and rephrased.
- (L33) This is very nitpicky but PyTorch is typically stylized as "PyTorch" or `torch`, not `pytorch`.
- (L39) `rpy2` should be cited (see e.g., https://joss.theoj.org/papers/10.21105/joss.00862.pdf)
- (Figure 1b) Please add units to the run times (i.e., "hetGPy (0.08s)", etc.)
- (L68) "Homoscedastic" is written here, but spelt with a "k" everywhere else. Either one is fine, but should be consistent.
- (L91) This citation does not render properly.
- (L104) "R" should be capitalized.
- (L172) The citation for Surjanovic & Bingham should have a year and URL, since it's a website. They suggest 2013 as the year on the website (https://www.sfu.ca/~ssurjano/about.html).
