+++ paper/analysis/table.tex
@@ -1,25 +1,31 @@
 \documentclass{standalone}
 \usepackage{booktabs, multirow, amsmath}
 
+
 \begin{document}
+
 
 \begin{tabular}{llrrrrrr}
 \toprule
- &  & \multicolumn{3}{c}{Time (Seconds)} & \multicolumn{3}{c}{Number of Function Evaluations} \\ \cmidrule(lr){3-5} \cmidrule(lr){6-8}
+&  & \multicolumn{3}{c}{Time (Seconds)} & \multicolumn{3}{c}{Number of Function Evaluations} \\ \cmidrule(lr){3-5} \cmidrule(lr){6-8}
  Covariance & Experiment & GPyTorch (na\"ive) & GPyTorch (SK) & hetGPy & GPyTorch (na\"ive) & GPyTorch (SK) & hetGPy \\
 \midrule
-\multirow[t]{5}{*}{Gaussian} & Branin & 98.45 & 5.62 & 2.69 & 13 & 18 & 13 \\
- & Goldstein-Price & 119.02 & 3.69 & 1.90 & 21 & 12 & 10 \\
- & Hartmann-4D & 98.57 & 4.58 & 4.94 & 18 & 15 & 16 \\
- & Hartmann-6D & 466.38 & 19.85 & 21.47 & 95 & 69 & 40 \\
- & Sphere-6D & 206.05 & 13.99 & 7.70 & 42 & 51 & 20 \\
+\multirow[t]{5}{*}{Gaussian} & Branin & 98.45 & 5.61 & 2.96 & 13 & 18 & 13 \\
+ & Goldstein-Price & 119.02 & 3.7 & 2.04 & 21 & 12 & 10 \\
+ & Hartmann-4D & 98.57 & 4.67 & 5.15 & 18 & 15 & 16 \\
+ & Hartmann-6D & 466.38 & 21.05 & 21.49 & 95 & 69 & 40 \\
+ & Sphere-6D & 206.05 & 14.54 & 7.83 & 42 & 51 & 20 \\
 \cline{1-8}
-\multirow[t]{5}{*}{Matern $\nu = 5/2$} & Branin & 90.501 & 4.765 & 6.563 & 13 & 12 & 20 \\
- & Goldstein-Price & 153.03 & 4.41 & 4.01 & 20 & 11 & 15 \\
- & Hartmann-4D & 201.21 & 6.90 & 6.76 & 21 & 18 & 16 \\
- & Hartmann-6D & 339.39 & 28.43 & 45.23 & 53 & 67 & 57 \\
- & Sphere-6D & 147.12 & 16.03 & 16.35 & 24 & 38 & 14 \\
+\multirow[t]{5}{*}{Matern $\nu = 5/2$} & Branin & 90.5 & 4.88 & 6.62 & 13 & 12 & 20 \\
+ & Goldstein-Price & 153.03 & 4.55 & 4.0 & 20 & 11 & 15 \\
+ & Hartmann-4D & 201.21 & 7.24 & 6.73 & 21 & 18 & 16 \\
+ & Hartmann-6D & 339.39 & 29.45 & 45.78 & 53 & 67 & 57 \\
+ & Sphere-6D & 147.12 & 16.89 & 16.46 & 24 & 38 & 14 \\
+\cline{1-8}
 \bottomrule
 \end{tabular}
+
+
+
 
 \end{document}
