### 2024-12-07 742bd3e
update affiliations
