### 2024-12-07 davidogara/hetGPy#15 @davidogara
Thanks for the suggestions @DanWaxman -- for the GPyTorch and homGP experiments, I ended up running them 5 times and reported the median (runtime in seconds and number of function evaluations). I also added the computational resources in the paper (specifically, 2019 MacBook Pro with a 2.6 GHz 6-Core Intel Core i7 processor, 16GB RAM, macOS Sonoma 14.7.1)
