### 2024-12-07 8de1a04
first round of paper edits
