+++ paper/paper.md
@@ -37,24 +37,24 @@ bibliography: paper.bib
 ---
 
 # Summary
-Computer experiments are ubiquitous in the physical and social sciences. When experiments are time-consuming to run, emulator (or surrogate) models are often used to map the input–output response surface of the simulator, treating it as a black box. The workhorse function of emulator models is Gaussian process regression (GPR). GPs provide flexible, non-linear regression targets with good interpolation properties and uncertainty quantification. However, it is well-known that naïve GPR scales cubically with input size, and specifically involves intensive computation of matrix determinants and solving linear systems [@garnett_bayesian_2023;@gramacy_surrogates_2020] when fitting hyperparameters. Further, naïve GPR with noisy observations typically assumes an independent, identically-distributed noise process, but many data-generating mechanisms, especially those found in stochastic computer simulation, exhibit input-dependent noise (also known as heteroskedasticity) [@baker_analyzing_2020]. The software package `hetGP` [@binois_hetgp_2021] alleviates these both of these concerns: when the dataset of interest contains replicates, as it is possible to perform inference and prediction with cost growing cubically in the number of unique design locations n rather than the full dataset of size N and can jointly model the mean and noise process as two coupled GPs, allowing smooth noise dynamics over parameter space [@binois_practical_2018]. The package has been used in a variety of contexts, such as statistics [@binois_practical_2018;@binois_replication_2019], biology [@lazaridis_genetic_2022] and computational epidemiology [@shattock_impact_2022]. We present a Python reimplementation `hetGPy`, developed in part due to Python’s widespread use in computer simulation [@downey_modeling_2023;@kinser_modeling_2022]. 
+Computer experiments are ubiquitous in the physical and social sciences. When experiments are time-consuming to run, emulator (or surrogate) models are often used to map the input–output response surface of the simulator, treating it as a black box. The workhorse function of emulator models is Gaussian process regression (GPR). GPs provide flexible, non-linear regression targets with good interpolation properties and uncertainty quantification. However, it is well-known that naïve GPR scales cubically with input size, and specifically involves intensive computation of matrix determinants and solving linear systems when fitting hyperparameters [@garnett_bayesian_2023;@gramacy_surrogates_2020]. Further, naïve GPR with noisy observations typically assumes an independent, identically-distributed noise process, but many data-generating mechanisms, especially those found in stochastic computer simulation, exhibit non-equal variance noise (also known as heteroskedasticity) [@baker_analyzing_2020]. The software package `hetGP` [@binois_hetgp_2021] alleviates both of these concerns: when the dataset of interest contains multiple observations per location, as it is possible to perform inference and prediction with cost growing cubically in the number of unique design locations $n$ rather than the full dataset of size $N$[@binois_practical_2018]. Further, `hetGP` also models the mean and noise processes as two coupled GPs, allowing smooth noise dynamics over parameter space [@binois_practical_2018;@gramacy_surrogates_2020]. The package has been used in a variety of contexts, such as statistics [@binois_practical_2018;@binois_replication_2019], biology [@lazaridis_genetic_2022] and computational epidemiology [@shattock_impact_2022]. We present a Python reimplementation `hetGPy`, developed in part due to Python’s widespread use in computer simulation [@downey_modeling_2023;@kinser_modeling_2022]. 
 
 # Statement of Need
 
[…]
