+++ paper/paper.bib
@@ -16,7 +16,7 @@
 }
 
 @article{binois_hetgp_2021,
-	title = {\textbf{{hetGP}} : {Heteroskedastic} {Gaussian} {Process} {Modeling} and {Sequential} {Design} in \textit{{R}}},
+	title = {\textbf{{hetGP}} : {Heteroskedastic} {Gaussian} {Process} {Modeling} and {Sequential} {Design} in {R}},
 	volume = {98},
 	issn = {1548-7660},
 	shorttitle = {\textbf{{hetGP}}},
@@ -47,8 +47,14 @@
 	file = {O'Gara et al_2023_TRACE‐Omicron.pdf:/Users/davidogara/Library/CloudStorage/Box-Box/Zotero/O'Gara et al_2023_TRACE‐Omicron.pdf:application/pdf},
 }
 
+@article{Gautier:2008,
+  title={rpy2: A Simple and Efficient Access to R from Python},
+  author={Gautier, L},
+  journal={URL http://rpy. sourceforge. net/rpy2. html},
+  year={2008}
+}
 @article{aylett-bullock_j_2021,
-	title = {J {\textless}span style="font-variant:small-caps;"{\textgreater}une{\textless}/span{\textgreater} : open-source individual-based epidemiology simulation},
+	title = {JUNE: open-source individual-based epidemiology simulation},
 	volume = {8},
 	issn = {2054-5703},
 	shorttitle = {J {\textless}span style="font-variant},
@@ -355,6 +361,8 @@ Publisher: Nature Publishing Group},
 @misc{surjanovic_virtual_nodate,
 	title = {Virtual {Library} of {Simulation} {Experiments}: {Test} {Functions} and {Datasets}},
 	author = {Surjanovic, S. and Bingham, D.},
+	year = {2013},
+	url = {https://www.sfu.ca/~ssurjano/},
 	note = {Published: Retrieved March 15, 2024, from http://www.sfu.ca/ ssurjano},
 }
 
