### 2024-12-08 davidogara/hetGPy#17 @DanWaxman
`RuntimeWarning` in tests
All tests pass, but I have various `RuntimeWarning`s thrown. These are the relevant ones:

> tests/test_changing_dimensions.py::test_het
>   /path/to/hetGPy/hetgpy/hetGP.py:909: RuntimeWarning: invalid value encountered in scalar divide
>     init['g'] = mean_var_replicates_nugs / np.var(nugs_est0,ddof=1)
>
> tests/test_changing_dimensions.py::test_het
>   /path/to/hetGPy/hetgpy/homGP.py:51: RuntimeWarning: divide by zero encountered in log
>     loglik = (-N / 2.0) * np.log(2*np.pi) - (N / 2.0) * np.log(psi) + (1.0 / 2.0) * ldetKi - (N - n)/2.0 * np.log(g) - (1.0 / 2.0) * np.log(mult).sum() - (N / 2.0)
>
> tests/test_changing_dimensions.py::test_het
>   /path/to/hetGPy/hetgpy/homGP.py:86: RuntimeWarning: invalid value encountered in scalar divide
>     tmp1[i] = (k/2 * (KiZ0.T @ dC_dthetak) @ KiZ0 /(((Z.T @ Z) - (Z0 * mult).T @ Z0)/g + psi) - 1/2 * np.trace(Ki @ dC_dthetak)).squeeze() # replaces trace_sym
>
> tests/test_changing_dimensions.py::test_het
>   /path/to/hetGPy/hetgpy/homGP.py:89: RuntimeWarning: invalid value encountered in scalar divide
>     tmp2 = k/2 * ((Z.T @ Z - (Z0 * mult).T @ Z0)/g**2 + np.sum(KiZ0**2/mult)) / ((Z.T @ Z - (Z0 * mult).T @ Z0)/g + psi) - (k - n)/ (2*g) - 1/2 * np.sum(np.diag(Ki)/mult)
>
> tests/test_changing_dimensions.py::test_het
>   /path/to/hetGPy/hetgpy/homGP.py:389: RuntimeWarning: divide by zero encountered in divide
>     self['Ki'] /= self['nu_hat'] # this is a subtle difference between R and Python.
>
> tests/test_changing_dimensions.py::test_het
>   /path/to/hetGPy/hetgpy/homGP.py:392: RuntimeWarning: invalid value encountered in matmul
>     mean = self['beta0'] + kx @ (self['Ki'] @ (self['Z0'] - self['beta0']))
>
> tests/test_changing_dimensions.py::test_het
>   /path/to/hetGPy/hetgpy/homGP.py:397: RuntimeWarning: invalid value encountered in matmul
>     sd2 = self['nu_hat'] - np.diag(kx @ ((self['Ki'] @ kx.T))) + (1- (self['Ki'].sum(axis=0))@ kx.T)**2/self['Ki'].sum()
>
> tests/test_changing_dimensions.py::test_het
>   /path/to/hetGPy/.venv/lib/python3.10/site-packages/numpy/_core/_methods.py:53: RuntimeWarning: invalid value encountered in reduce
>     return umr_sum(a, axis, dtype, out, keepdims, initial, where)
>
> tests/test_changing_dimensions.py::test_het
>   /path/to/hetGPy/hetgpy/homGP.py:421: RuntimeWarning: invalid value encountered in multiply
>     self['Ki']*=self['nu_hat']
>
> tests/test_optim_and_update/test_update_hetGP.py::test_update
>   /path/to/hetGPy/hetgpy/hetGP.py:143: RuntimeWarning: invalid value encountered in log
>     loglik = -N/2 * np.log(2*np.pi) - N/2 * np.log(psi) + 1/2 * ldetKi - 1/2 * np.sum((mult - 1) * np.log(Lambda) + np.log(mult)) - N/2
>
> tests/test_optim_and_update/test_update_hetGP.py::test_update
>   /path/to/hetGPy/hetgpy/hetGP.py:300: RuntimeWarning: invalid value encountered in log
>     loglik = -N/2 * np.log(2*np.pi) - N/2 * np.log(psi/N)  + 1/2 * ldetKi - 1/2 * np.sum((mult - 1) * np.log(Lambda) + np.log(mult)) - N/2

It seems like these are known to occur -- after some debugging, I realized that these are handled

https://github.com/davidogara/hetGPy/blob/8de1a04a78765014dca0021731fcf32ede8a19f1/hetgpy/homGP.py#L322-L324

and pop up in the stack because of calls to `mleHomGP` from `mleHetGP`

https://github.com/davidogara/hetGPy/blob/8de1a04a78765014dca0021731fcf32ede8a19f1/hetgpy/hetGP.py#L915-L921

Even though these don't seem to affect the results (in the sense that all tests pass), users should not get a `RuntimeWarning` about arithmetic in expected behavior.

Part of https://github.com/openjournals/joss-reviews/issues/7518
