### 2024-12-08 davidogara/hetGPy#15 @DanWaxman
Thanks for the speedy reply, I'll close this issue out!
