### 2024-12-08 davidogara/hetGPy#18 @DanWaxman
Tests Without Comparisons
Some tests run the R code and hetGPy code, but never actually compare the results as far as I can tell.

For example, in `tests/test_optim_and_update/test_update_homGP.py`, there *is* such a check:
https://github.com/davidogara/hetGPy/blob/8de1a04a78765014dca0021731fcf32ede8a19f1/tests/test_optim_and_update/test_update_homGP.py#L62
but in the corresponding test case for `hetGP`, there is no assertion statement:
https://github.com/davidogara/hetGPy/blob/8de1a04a78765014dca0021731fcf32ede8a19f1/tests/test_optim_and_update/test_update_hetGP.py#L8-L62

This seems to be a common problem in the `test_optim_and_update` folder. For example,
https://github.com/davidogara/hetGPy/blob/8de1a04a78765014dca0021731fcf32ede8a19f1/tests/test_optim_and_update/test_IMSPE_optim.py#L8-L53
also does not have a test.

Looking briefly, this does not seem to be an issue in the `tests` folder.

Part of https://github.com/openjournals/joss-reviews/issues/7518
