### 2024-12-09 a6eae95
wrap minimize calls in np.errstate context
