### 2024-12-09 davidogara/hetGPy#17 @davidogara
@DanWaxman thanks for pointing this out. I think the best way to handle these would be wrapping the `scipy.optimize.minimize` call in a `np.errstate(divide='ignore',invalid='ignore')`. I set this up to be turned on or off in `homGP.mleHomGP` and `hetGP.mleHetGP` with an extra argument to the `settings` dictionary

You can see in `tests/test_changing_dimensions.py`: https://github.com/davidogara/hetGPy/blob/main/tests/test_changing_dimensions.py#L9

Setting `ignore_MLE_divide_invalid=True` correctly suppresses the runtime warnings (and they get properly handled by the rest of the optimization routine), but if someone wanted to see them, they could set that option to False

Let me know if you have any thoughts on the above, or if you have any suggestions on best practice for handling this kind of thing
