### 2024-12-09 ac3258d
errstate context manager only for homGP calls within hetGP
