### 2024-12-09 davidogara/hetGPy#17 @DanWaxman
Yes, thanks for clarifying. This looks good to me!
