### 2024-12-09 1cc2640
lower anisotropic theta bound to avoid warning
