### 2024-12-09 8456992
add sequential EI test
