### 2024-12-09 davidogara/hetGPy#18 @davidogara
Thanks for bringing this up @DanWaxman -- I think that these are residual from an earlier version of the library where the sequential routines (so the `update` methods and any of the acquisition functions) had bugs when you tried to run them end-to-end, but these have been resolved now.

I think a more appropriate test would be to consolidate these tests into one that confirms that the sequential function acquisitions from `homGP` (in R) and `homGP` (in hetgpy) for a simple example are the same.

I added the test:
https://github.com/davidogara/hetGPy/blob/main/tests/test_optim_and_update/test_sequential_EI.py

Which passes. Given that more complicated examples might be more tricky to test, I'd be inclined to have this test included in the testing suite and remove `test_IMSPE_optim`, `test_crit_optim`, `test_update_hetGP`, and `test_update_homGP`

If you think that is appropriate, let me know. Thanks again for the careful review
