### 2024-12-10 davidogara/hetGPy#18 @DanWaxman
I think there's probably value in keeping these "trickier" tests -- e.g., if they were useful to help identify and resolve old bugs in `update` methods, they may well prove useful in catching some potential future bugs. But I don't view this as critical, the proposed `test_sequential_EI.py` should be sufficient; I just think broader coverage rarely hurts. Thanks again for being so receptive and timely in your replies.
