### 2024-12-10 davidogara/hetGPy#20 @DanWaxman
Documentation and Potential Typing
I think the last remaining aspect of my review is documentation, with some further documentation in order in some places. In particular, some functions which are documented in `hetGP`are not documented in `homGP`, e.g. `logLikHom`: https://github.com/davidogara/hetGPy/blob/8456992bad244d2ece4720e9eba93807b7812086/hetgpy/homGP.py#L31

I understand some methods are not API level, but e.g., `predict` and `update` should be, and are not documented. In my opinion, full documentation is necessary in cases like this to satisfy the "Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?" part of the checklist.

While I do not find type hints entirely necessary, I believe it would be a big plus. A pure `numpy` solution would be using [`numpy.typing`](https://numpy.org/doc/stable/reference/typing.html). Even better would be [`jaxtyping`](https://github.com/patrick-kidger/jaxtyping), which contrary to the name is fully compatible with regular `numpy`, and allows for array shapes and static type checking.

Part of https://github.com/openjournals/joss-reviews/issues/7518
