### 2024-12-10 davidogara/hetGPy#14 @DanWaxman
Also, as a very small additional comment, it is mentioned that "[hetGPy] is well-posed to engage with the wider Python community for future extension such as arbitrary kernel functions with auto-differentiation methods facilitated by PyTorch." It probably makes sense here to mention `jax` rather than `numpy`, since `jax.numpy` implements (with caveats) the majority of the `numpy` API directly.
