### 2024-12-10 cc8bd5f
add update tests with 2 steps total
