### 2024-12-10 347d8e7
add test for crit_optim
