### 2024-12-10 davidogara/hetGPy#18 @davidogara
That makes sense-- I was able to simply the tests a bit (the results are far more stable when we only do a single update step in the acquisition functions) and compare the results between hetGP and hetgpy (and they pass)

For a future version of the library I think it would make sense to have a longer worked example test that does more update steps (say like 5 or even 10) that shows that hetGP (from R) and hetgpy stay in synch
