### 2024-12-10 davidogara/hetGPy#20 @davidogara
@DanWaxman thanks for the suggestion, I think these would be very helpful too.

I'll take a pass at updating the log-likelihood, predict, and update methods, as well as adding type hints to those (and the MLE methods)
