### 2024-12-10 af509c9
remove env from LL
