### 2024-12-10 0cbfe94
update predict and rebuild docstrings
