### 2024-12-10 e91c0e6
docstring for gradient
