### 2024-12-10 0957468
add strip; update docstrings
