+++ hetgpy/hetGP.py
@@ -1481,6 +1481,48 @@ class hetGP:
     
     def update(self,Xnew, Znew, ginit = 1e-2, lower = None, upper = None, noiseControl = None, settings = None,
                          known = {}, maxit = 100, method = 'quick'):
+        r'''
+        Update model object with new observations
+
+        Parameters
+        ----------
+        Xnew: ndarray_like
+            new inputs (one observation per row)
+        Znew: ndarray_like
+            new observations
+        lower: ndarray_like
+            lower bound for lengthscales. If not provided extracted from `self`
+        upper: ndarray_like
+            upper bound for lengthscales. If not provided extracted from `self`
+        noiseControl: dict
+            noise bounds. If not provided extracted from `self`
+        settings: dict
+            options for optimization. If not provided extracted from `self`
+        known: dict
+            known hyperparameters to fix at values
+        maxit: int
+            maximum number of iterations for hyperparameter optimization
+        
+        Returns
+        -------
+        self, potentially with updated hyperparameters
+        
+
+        Notes
+        -----
+        If hyperparameters do not need to be updated, maxit can be set to 0. In this case it is possible to pass NAs in Znew, then the model can still be used to provide updated variance predictions
+        
+        Example
+        -------
+        >>> import numpy as np, from hetgpy import hetGP
+        >>> X = np.array([1, 2, 3, 4, 12]).reshape(-1,1)
+        >>> Y = np.array([0, -1.75, -2, -0.5, 5])
+        >>> model = hetGP().mle(X,Y,known={'theta':np.array([10.0]),'g':1e-8})
+        >>> Xnew  = np.array([2.5]).reshape(-1,1)
+        >>> Znew  = model.predict(Xnew)['mean']
+        >>> model.update(Xnew,Znew)
+        '''
+        
         # first reduce Xnew/Znew in case of potential replicates
         newdata = find_reps(Xnew, Znew, normalize = False, rescale = False)
   
@@ -1632,6 +1674,33 @@ class hetGP:
 
         newmodel = deepcopy(self)
         return newmodel
+    def rebuild(self, robust = False):
+        r'''
+        Rebuilds inverse covariance matrix in homGP object (usually after saving). Works in tandem with `strip`
+
+        Parameters
+        ----------
+
+        robust: bool
+            use `np.linalg.pinv` for covariance matrix inversion. Otherwise use cholesky
+        
+        Returns
+        -------
+        self with rebuilt inverse covariance matrix
+        '''
+        if robust :
+            self['Ki'] <- np.linalg.pinv(
+                cov_gen(X1 = self['X0'], theta = self['theta'], type = self['covtype']) + np.diag(self['eps'] + self['g'] / self['mult'])
+            ).T
+            self['Ki'] /= self['nu_hat']
+        else:
+            ki = np.linalg.cholesky(
+            cov_gen(X1 = self['X0'], theta = self['theta'], type = self['covtype']) + np.diag(self['eps'] + self['g'] / self['mult'])
+            ).T
+            ki = dtrtri(ki)[0]
+            self['Ki'] = ki @ ki.T
+        return self
+
     def LOO_preds_nugs(self, i):
         '''
         Leave-out-out predictions for the nugget
@@ -1656,6 +1725,19 @@ class hetGP:
     
     
     def plot(self,type='diagnostics',**kwargs):
+        r'''
+        Plot model behavior. See hetgpy.plot for more details on types of plots
+
+        Parameters
+        ----------
+        type: str
+            type of plot: currently supported are "diagnostics" and "iterates"
+
+
+        Returns
+        -------
+        fig, ax: figure and axis object
+        '''
         ptypes = ('diagnostics','iterates')
[…]
