### 2024-12-16 3561255
update strip and rebuild methods
