### 2024-12-16 4e296d5
add typehints
