### 2024-12-16 f9112e4
update typehints
