### 2024-12-16 davidogara/hetGPy#20 @davidogara
@DanWaxman just updating this thread to let you know I added documentation and type hints to the likelihoods, gradient, MLE, update, and predict methods. Please feel free to take a look and let me know what you think!
