### 2024-12-16 davidogara/hetGPy#13 @davidogara
@Edenhofer just checking in to see if you had any feedback on the analysis script, or if not, if you wouldn't mind closing this issue. Happy to address any follow-up if you have it!
