### 2024-12-16 davidogara/hetGPy#13 @Edenhofer
I'm happy with the script. Feel free to close the issue. As I didn't open the issue, I'm unable to close it myself.
