+++ hetgpy/hetGP.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import numpy as np
 import warnings
 from time import time
