+++ hetgpy/homGP.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import numpy as np
 from itertools import chain
 import warnings
