### 2024-12-16 2eb8158
add np. to log and exp
