+++ paper/paper.bib
@@ -509,3 +509,20 @@ Publisher: Nature Publishing Group},
 	year = {1995},
 	Bdsk-Url-1 = {https://doi.org/10.1137/0916069}
 }
+
+
+@book{harville_matrix_1997,
+	address = {New York, NY},
+	author = {Harville, David A.},
+	publisher = {Springer},
+	title = {Matrix {Algebra} {From} a {Statistician}'s {Perspective}},
+	year = {1997}}
+
+@software{jax2018github,
+  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url = {http://github.com/jax-ml/jax},
+  version = {0.3.13},
+  year = {2018},
+}
+
