### 2024-12-16 davidogara/hetGPy#14 @davidogara
@Edenhofer @DanWaxman  I've gotten a address your comments on the paper. I've repasted the more substantive ones (many of them I just addressed by fixing typos). Since many of your comments overlapped I decided to address them at the same time, but I'd be happy to break these into separate comments

For @Edenhofer:
- l21ff IMO the sentence is a bit overloaded and hard to read. Please rephrase.
   - Thanks, I re-worked this and the following sentences to hopefully be more clear, but please let me know if it is not
- l23 here and in the following, please be more specific with what you mean with “replicas”/“replication”. I think you mean multiple observations at the same location. I would suggest spelling this out instead of using the term “replication”.
   - Thanks for mentioning this: I re-worked this sentence to more specifically spell out what I mean by replicates (and you are correct, that it is multiple observations at the same input location)
- l46 the default behavior of a library is not too relevant. Rather say whether these libraries can do that. It’s been a while since I’ve - used them but AFAIK some of these libraries support learning the mean and variance.
  - Thanks, I re-worded to mention
- l49 Could the authors briefly elaborate on why hetGPy is more computationally efficient.
  - Yes, I made it more explicit in the text (that hetGPy solves a problem of size n=1000 rather than N=~5000 in these examples), which combined with the explanation around line 20, should make things more clear

And for @DanWaxman:

- (L21-L25) I also think this sentence should be broken up into several sentences and rephrased.
   - Thanks, I re-worked this section by breaking it up into several sentences which should hopefully help the explanation

Thank you both again for your feedback on the paper. Please let me know if you have any follow-up and I'd be happy to address it.
