### 2024-12-17 davidogara/hetGPy#14 @Edenhofer
Thanks for the update. Could you please resolve the following and share a link to the updated PDF ? https://github.com/openjournals/joss-reviews/issues/7518#issuecomment-2548222082
