### 2024-12-17 davidogara/hetGPy#14 @davidogara
Thank you! You should be able to download the paper proof from this link:
https://github.com/openjournals/joss-reviews/issues/7518#issuecomment-2548889664
