### 2024-12-17 davidogara/hetGPy#14 @DanWaxman
Thanks, these changes look good! As a remaining typo, on (L43) there should be a space after "rpy2".
