### 2024-12-17 davidogara/hetGPy#21 @DanWaxman
Deployment Failing
As a final point to https://github.com/openjournals/joss-reviews/issues/7518, the deployments to PyPi are failing. The logs are expired, so I am not sure why (the warnings look innocuous) -- but this should be resolved.
