### 2024-12-17 davidogara/hetGPy#18 @DanWaxman
Thanks!

> For a future version of the library I think it would make sense to have a longer worked example test that does more update steps (say like 5 or even 10) that shows that hetGP (from R) and hetgpy stay in synch

Perhaps it makes sense to create a separate GH issue for this in case a future contributor to the library decides to take this on. But I'll close this issue for now.
