+++ paper/paper.md
@@ -41,7 +41,7 @@ Computer experiments are ubiquitous in the physical and social sciences. When ex
 
 # Statement of Need
 
-Python is a popular language for software development, data science, and computer experimentation. Its object orientated framework, high-level functionality, and third-party libraries such as `numpy` [@harris_array_2020], `scipy` [@virtanen_scipy_2020], `PyTorch` [@paszke_pytorch_2019] and `scikit-learn` [@pedregosa_scikit-learn_2011] make it a powerful tool for academic and industry professionals alike. Python is especially popular for computer simulation, with one particular example being the widespread use of Python models of COVID-19 spread [@aylett-bullock_j_2021;@kerr_covasim_2021;@ogara_traceomicron_2023]. `hetGPy` is well-posed for sequential design of Python models, mirroring the functionality of hetGP without having to rely on intermediate libraries such as `reticulate` [@ushey_reticulate_2023] or `rpy2`[@Gautier:2008], or in a more laborious case, converting a Python simulation to R. 
+Python is a popular language for software development, data science, and computer experimentation. Its object orientated framework, high-level functionality, and third-party libraries such as `numpy` [@harris_array_2020], `scipy` [@virtanen_scipy_2020], `PyTorch` [@paszke_pytorch_2019] and `scikit-learn` [@pedregosa_scikit-learn_2011] make it a powerful tool for academic and industry professionals alike. Python is especially popular for computer simulation, with one particular example being the widespread use of Python models of COVID-19 spread [@aylett-bullock_j_2021;@kerr_covasim_2021;@ogara_traceomicron_2023]. `hetGPy` is well-posed for sequential design of Python models, mirroring the functionality of hetGP without having to rely on intermediate libraries such as `reticulate` [@ushey_reticulate_2023] or `rpy2` [@Gautier:2008], or in a more laborious case, converting a Python simulation to R. 
 
 The state of the art for GPR in Python is `GPyTorch` [@gardner_gpytorch_2018], facilitated by black-box matrix–matrix multiplication (BBMM) which is extremely computationally efficient on GPUs. Other GPR routines for Python exist as well and can be found in libraries such as `PyMC` [@abril-pla_pymc_2023], `GPflow` [@matthews_gpflow_2017], and `GPJax` [@pinder_gpjax_2022]. However, while it is possible to jointly model the mean and variance as coupled GPs with these libraries, they do not take advantage of replication in datasets, meaning that under large degrees of replication and input-dependent noise, as is common in stochastic computer experiments [@baker_analyzing_2020], `hetGPy` will be more computationally efficient. \autoref{fig:hetgpy-example} (a) shows a heteroskedastic GP fit to a simulated motorcycle accident dataset [@silverman_aspects_1985]. While it is possible to model input-dependent noise in `GPyTorch` or `BoTorch` [@balandat_botorch_2020], specifying a smooth noise process in the method of [@binois_practical_2018] would require a custom implementation. \autoref{fig:hetgpy-example} (b) illustrates the results of a simple one-dimensional example where we compare the model fits using both `hetGPy` and `GPyTorch`. While both models result in similar predictions, `hetGPy` is far faster, performing exact inference in less than 1 second, while GPyTorch takes over 10 seconds.
 
