### 2024-12-17 davidogara/hetGPy#14 @davidogara
Thanks @DanWaxman! I addressed that last typo.
