### 2024-12-17 davidogara/hetGPy#14 @Edenhofer
Thanks for your work on refining the paper. I am happy with the changes and have no further comments :)
