### 2024-12-18 #7518 @DanWaxman
I'm also happy with the submission -- I recommend the paper for publication with the resolution of https://github.com/davidogara/hetGPy/issues/21 (which I don't imagine will be a big deal). Thank you to the authors for being so receptive and responsive to feedback!
