### 2024-12-18 #7518 @matthewfeickert
@davidogara can you please update us here when https://github.com/davidogara/hetGPy/issues/21 is resolved?
