### 2024-12-18 davidogara/hetGPy#14 @davidogara
Thank you both! In that case, I will close the issue.
