+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "hetgpy"
-version = "0.1.post1"
+version = "1.0.0"
 description = "Python implementation of hetGP"
 authors = [
 {name="David O'Gara", email="david.ogara@wustl.edu"},
