+++ .github/workflows/release-to-pypi.yaml
@@ -41,7 +41,7 @@ jobs:
     - build
     runs-on: ubuntu-20.04
     environment:
-      name: release
+      name: pypi
       url: https://pypi.org/p/hetgpy  # Replace <package-name> with your PyPI project name
     permissions:
       id-token: write  # IMPORTANT: mandatory for trusted publishing
@@ -104,7 +104,7 @@ jobs:
     runs-on: ubuntu-20.04
 
     environment:
-      name: release
+      name: testpypi
       url: https://test.pypi.org/p/hetgpy
 
     permissions:
