+++ .github/workflows/release-to-pypi.yaml
@@ -25,8 +25,8 @@ jobs:
       run: git submodule update --init --recursive
     - name: symlink eigen to /include
       run: sudo cp -r eigen/Eigen /usr/local/include/
-    - name: Build a binary wheel and a source tarball
-      run: python3 -m build
+    - name: Build a source tarball
+      run: python3 -m build --sdist
     - name: Store the distribution packages
       uses: actions/upload-artifact@v3
       with:
