+++ .github/workflows/release-to-pypi.yaml
@@ -5,7 +5,6 @@ on: push
 jobs:
   build:
     name: Build distribution 📦
-    if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
     runs-on: ubuntu-20.04
 
     steps:
@@ -61,7 +60,7 @@ jobs:
       Sign the Python 🐍 distribution 📦 with Sigstore
       and upload them to GitHub Release
     needs:
-    - publish-to-pypi
+    - build
     runs-on: ubuntu-20.04
 
     permissions:
