+++ .github/workflows/release-to-pypi.yaml
@@ -74,11 +74,9 @@ jobs:
         name: python-package-distributions
         path: dist/
     - name: Sign the dists with Sigstore
-      uses: sigstore/gh-action-sigstore-python@v2.1.1
+      uses: sigstore/gh-action-sigstore-python@v3.0.0
       with:
-        inputs: >-
-          ./dist/*.tar.gz
-          ./dist/*.whl
+        inputs: ./dist/*.tar.gz
     - name: Create GitHub Release
       env:
         GITHUB_TOKEN: ${{ github.token }}
