### 2024-12-18 7eeb5bd
sigstore python 3.10
