+++ .github/workflows/release-to-pypi.yaml
@@ -73,11 +73,10 @@ jobs:
       with:
         name: python-package-distributions
         path: dist/
-    - name: View files
-      run: ls
     - name: Sign the dists with Sigstore
       uses: sigstore/gh-action-sigstore-python@v3.0.0
       with:
+        python-version: "3.10"
         inputs: ./dist/*.tar.gz
     - name: Create GitHub Release
       env:
