+++ .github/workflows/release-to-pypi.yaml
@@ -74,9 +74,11 @@ jobs:
         name: python-package-distributions
         path: dist/
     - name: Sign the dists with Sigstore
+      uses: actions/setup-python@v5
+        with:
+        python-version: '3.9'    
       uses: sigstore/gh-action-sigstore-python@v3.0.0
       with:
-        python-version: "3.10"
         inputs: ./dist/*.tar.gz
     - name: Create GitHub Release
       env:
