+++ .github/workflows/release-to-pypi.yaml
@@ -73,10 +73,12 @@ jobs:
       with:
         name: python-package-distributions
         path: dist/
-    - name: Sign the dists with Sigstore
+    - name: Install Python
       uses: actions/setup-python@v5
-        with:
-        python-version: '3.9'    
+      with:
+        python-version: '3.9'
+    - run: python3 --V
+    - name: Sign with sigstore
       uses: sigstore/gh-action-sigstore-python@v3.0.0
       with:
         inputs: ./dist/*.tar.gz
