### 2024-12-18 3bdf874
check python version
