+++ .github/workflows/release-to-pypi.yaml
@@ -77,7 +77,7 @@ jobs:
       uses: actions/setup-python@v5
       with:
         python-version: '3.9'
-    - run: python --V
+    - run: python -V
     - name: Sign with sigstore
       uses: sigstore/gh-action-sigstore-python@v3.0.0
       with:
