### 2024-12-18 davidogara/hetGPy#21 @davidogara
@DanWaxman I believe I have addressed this. There were some old environments I had to delete, but the workflow successfully pushed hetgpy v1.0 to pypi

Can you take a look and let me know what you think?
