### 2024-12-18 davidogara/hetGPy#21 @DanWaxman
Looks good to me! I confirmed in a Colab notebook that I can install `hetgpy-1.0.0` and run the motorcycle example.
