### 2024-12-18 davidogara/hetGPy#22 @matthewfeickert
[JOSS Review] Missing documentation components and Python version tests
Related to JOSS review issue https://github.com/openjournals/joss-reviews/issues/7518

The `hetGPy` docs currently (as of b345c6d9e37920f179ab26849cccd999f7edd8f0) don't include any installation instructions. The landing page just dumps you out onto either examples or the API reference.

Note from the [JOSS review checklist documentation section](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation) that the documentation should contain an overview/statement of need (or similar) and installation instructions (and the other things listed).

Also please fix the RST hyperlink typos in the examples. ([Example](https://hetgpy.readthedocs.io/en/latest/examples/00-hetGPy-Intro.html#Why-is-a-Python-port-needed-when-we-have-rpy2-to-call-R-code-from-Python?))

Related (but **not required**) request: Please consider enabling [versioned docs on ReadTheDocs](https://docs.readthedocs.io/en/stable/versions.html#versions) so you have something beyond `latest` and `stable`. This is especially relevant for people to get correct API reference documentation.

Additionally, related to testing, `hetGPy` currently requires Python 3.10+

https://github.com/davidogara/hetGPy/blob/b345c6d9e37920f179ab26849cccd999f7edd8f0/pyproject.toml#L10

but the tests test only `3.10` and not `3.10` and newer

https://github.com/davidogara/hetGPy/blob/b345c6d9e37920f179ab26849cccd999f7edd8f0/.github/workflows/run-tests.yml#L21-L23

https://github.com/davidogara/hetGPy/blob/b345c6d9e37920f179ab26849cccd999f7edd8f0/.github/workflows/run-tests-and-examples.yml#L21-L23

Unless there are clear reasons why testing on newer Pythons is known to fail / can't be done, the full supported Python range should be tested in CI.

The following isn't wrong, but just a note to simplify your life: The current build system uses

https://github.com/davidogara/hetGPy/blob/b345c6d9e37920f179ab26849cccd999f7edd8f0/pyproject.toml#L23-L25

but `wheel` isn't required here as `setuptools` will grab this as needed. So you can just use

```toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0", "setuptools-cpp","pybind11","requests"]
```

(c.f. https://learn.scientific-python.org/development/guides/packaging-simple/)

Also why do you need `requests` in there?
