### 2024-12-18 #7518 @matthewfeickert
Some (hopefully easy to address) things I noticed when doing a quick lookover before starting the post-review checklist: https://github.com/davidogara/hetGPy/issues/22
