### 2024-12-18 d8d6052
remove requests and wheel from build
