### 2024-12-18 4abdf7d
update run-tests to for multiple python versions
