+++ .github/workflows/run-tests-and-examples.yml
@@ -20,7 +20,7 @@ jobs:
       # Used to host cibuildwheel
       - uses: actions/setup-python@v5
         with:
-          python-version: '3.10' 
+          python-version: ['3.10', '3.11', '3.12', '3.13']
       - name: Install dependencies
         run: python -m pip install -r requirements.txt
       - name: Install dev dependencies
