+++ .github/workflows/release-to-pypi.yaml
@@ -59,6 +59,7 @@ jobs:
     name: >-
       Sign the Python 🐍 distribution 📦 with Sigstore
       and upload them to GitHub Release
+    if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes  
     needs:
     - build
     runs-on: ubuntu-20.04
