+++ .github/workflows/run-tests-and-examples.yml
@@ -19,7 +19,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       # Used to host cibuildwheel
-      - uses: actions/setup-python@v
+      - uses: actions/setup-python@v4
       - name: Install dependencies
         run: python -m pip install -r requirements.txt
       - name: Install dev dependencies
