+++ .github/workflows/run-tests-and-examples.yml
@@ -20,6 +20,8 @@ jobs:
       - uses: actions/checkout@v4
       # Used to host cibuildwheel
       - uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }
       - name: Install dependencies
         run: python -m pip install -r requirements.txt
       - name: Install dev dependencies
