+++ .github/workflows/run-tests.yml
@@ -14,17 +14,14 @@ jobs:
       fail-fast: false
       matrix:
         os: ['ubuntu-latest'] # [ubuntu-latest, windows-latest, macos-13, macos-14]
+        python-version: ['3.10','3.11','3.12','3.13']
 
     steps:
       - uses: actions/checkout@v4
       # Used to host cibuildwheel
       - uses: actions/setup-python@v5
         with:
-          python-version: |
-            3.10
-            3.11
-            3.12
-            3.13
+          python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: python -m pip install -r requirements.txt
       - name: Install dev dependencies
