### 2024-12-19 davidogara/hetGPy#22 @davidogara
Thanks for bringing these up @matthewfeickert. I was able to implement everything you suggested:

1. I added installation instructions to the documentation on readthedocs (https://hetgpy.readthedocs.io/en/latest/), as well as version tags (so you should see v1.0, stable, and latest)
2. I also added a short overview/statement of need to the documentation page
3. For the testing suite, I updated the testing workflow files so they run for python 3.10, 3.11, 3.12, and 3.13
4. And I updated the pyproject file (removed `wheel` and `requests`)

Thanks again for suggesting these, and let me know if you have any feedback on the above.
