### 2024-12-20 davidogara/hetGPy#22 @davidogara
@matthewfeickert apologies for the dual comment, but I noticed that I didn't address your comments about some the link typos in the example notebooks. Sorry about that, they are fixed and upload the readthedocs now
