### 2024-12-21 davidogara/hetGPy#22 @matthewfeickert
No worries @davidogara. You're addressing things super fast, which is excellent. I'm going to be a bit slower in getting to them given some professional deadlines and being a bit sick, but I hope to move things forward or given additional feedback before next week.
