### 2025-01-06 e0c1b5a
add jax covariance kernels
