+++ .github/workflows/run-tests.yml
@@ -13,10 +13,16 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ['ubuntu-20.04'] # [ubuntu-latest, windows-latest, macos-13, macos-14]
+        os: ['ubuntu-latest'] # [ubuntu-latest, windows-latest, macos-13, macos-14]
         python-version: ['3.10','3.11','3.12','3.13']
 
     steps:
+      - uses: actions/checkout@v4
+      - name: Setup R
+        uses: r-lib/actions/setup-r@v2
+        with:
+          r-version: '4.0'
+      - run: Rscript -e 'print("hello")'
       - uses: actions/checkout@v4
       # Used to host cibuildwheel
       - uses: actions/setup-python@v5
