+++ .github/workflows/run-tests.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Setup R
         uses: r-lib/actions/setup-r@v2
         with:
-          r-version: '4.0'
+          r-version: '4.3'
       - run: Rscript -e 'print("hello")'
       - uses: actions/checkout@v4
       # Used to host cibuildwheel
