+++ .github/workflows/run-tests-and-examples.yml
@@ -18,6 +18,12 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4
+      - name: Setup R
+        uses: r-lib/actions/setup-r@v2
+        with:
+          r-version: '4.3'
+      - run: Rscript -e 'print("hello")'
+      - uses: actions/checkout@v4
       # Used to host cibuildwheel
       - uses: actions/setup-python@v4
         with:
