### 2025-01-07 2a71a63
add experimental jax functionality
