+++ hetgpy/homGP_jax.py
@@ -1,0 +1,232 @@
+'''
+Implementation of homGP using jax
+
+**Note: currently experimental
+'''
+import jax
+jax.config.update("jax_enable_x64", True)
+from hetgpy import homGP
+from numpy.typing import ArrayLike, NDArray
+import jax.numpy as np
+from hetgpy.cov_jax import cov_gen_jax as cov_gen
+from jax.scipy.linalg import solve_triangular
+import warnings
+from time import time
+from hetgpy.utils import fast_tUY2
+from hetgpy.auto_bounds import auto_bounds
+from scipy import optimize
+NDArrayInt = NDArray[np.int_]
+MACHINE_DOUBLE_EPS = np.sqrt(np.finfo(float).eps)
+from jax.flatten_util import ravel_pytree
+
+class homGP_jax(homGP):
+    def __init__(self):
+        super().__init__()
+    def logLikHom(self,X0: ArrayLike, Z0: ArrayLike, Z: ArrayLike, mult: NDArrayInt, theta: ArrayLike, g: float, beta0: float | None = None, covtype: str = "Gaussian", eps: float = MACHINE_DOUBLE_EPS) -> float:
+        r'''
+        logLikHom but with autodiff
+        '''
+        n = X0.shape[0]
+        N = Z.shape[0]
+
+        C = cov_gen(X1 = X0, theta = theta, type = covtype)
+        self.C = C
+        Ki = np.linalg.cholesky(C + np.diag(eps + g / mult) ).T
+        ldetKi = - 2.0 * np.sum(np.log(np.diag(Ki)))
+        Ki = solve_triangular(Ki,np.eye(n)) #  -- equivalent of chol2inv -- see https://stackoverflow.com/questions/6042308/numpy-inverting-an-upper-triangular-matrix
+        Ki = Ki @ Ki.T     #  -- equivalent of chol2inv
+        
+        self.Ki = Ki
+        if beta0 is None:
+            beta0 = Ki.sum(axis=1) @ Z0 / Ki.sum()
+        self.beta0 = beta0
+
+        psi_0 = (Z0 - beta0).T @ Ki @ (Z0 - beta0)
+        psi = (1.0 / N) * ((((Z-beta0).T @ (Z-beta0) - ((Z0-beta0)*mult).T @ (Z0-beta0)) / g) + psi_0)
+        
+        # check for divide by zero warnings
+        #if psi <=0: return np.nan
+        #if g <= 0: return np.nan
+        loglik = (-N / 2.0) * np.log(2*np.pi) - (N / 2.0) * np.log(psi) + (1.0 / 2.0) * ldetKi - (N - n)/2.0 * np.log(g) - (1.0 / 2.0) * np.log(mult).sum() - (N / 2.0)
+            
+        
+        return loglik
+    
+    def mle_jax(self,X: ArrayLike, Z: ArrayLike, lower: ArrayLike | None = None, upper: ArrayLike | None = None, known: dict = dict(),
+                        noiseControl: dict = dict(g_bounds = (MACHINE_DOUBLE_EPS, 1e2)),
+                        init: dict = {},
+                        covtype: str = "Gaussian",
+                        maxit: int = 100, eps: float = MACHINE_DOUBLE_EPS, settings: dict = dict(returnKi = True, factr = 1e7)) -> None:
+        r'''
+        MLE, but auto-diffed
+        '''
+        known = known.copy()
+        init = init.copy()
+        if type(X) == dict:
+            X0 = X['X0']
+            Z0 = X['Z0']
+            mult = X['mult']
[…]
