### 2025-01-07 878ab71
remove 'pure' Python discussion in README
