+++ README.md
@@ -2,13 +2,14 @@
 
 `hetGPy` is a Python implementation of the `hetGP` R library.
 
-This package is designed to be a "pure" Python implementation of `hetGP`, with the goals of:
+This package has the goals of:
 *	Matching the behavior of the `R` package
-*	Having minimal dependencies, which are:
+*	Having minimal Python and external dependencies, which are:
     * `numpy` and `scipy` for computations
     * `matplotlib` for visualization
     * `joblib` for parallelization
     * `tqdm` for progress bars
+    * `Eigen` (C++) for fast calculations (when vectorization is non-obvious)
 
 The motivation for such a package is due to the rising popularity of implementing simulation models (also known as computer experiments) in Python. 
 
