+++ .github/workflows/ci.yaml
@@ -1,0 +1,47 @@
+name: Wheels
+
+on:
+  workflow_dispatch:
+jobs:
+
+  make_sdist:
+    name: Make SDist
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0 # Optional, use if you use setuptools_scm
+          submodules: true # Optional, use if you have submodules
+  
+      - name: Build SDist
+        run: pipx run build --sdist
+  
+      - uses: actions/upload-artifact@v4
+        with:
+          name: cibw-sdist
+          path: dist/*.tar.gz
+
+  build_wheels:
+    name: Wheel on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-13, macos-14]
+        python-version: ['3.10','3.11','3.12','3.13']
+  
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+          submodules: true
+  
+      - uses: pypa/cibuildwheel@v2.21
+        with:
+          ppython-version: ${{ matrix.python-version }}
+  
+      - name: Upload wheels
+        uses: actions/upload-artifact@v4
+        with:
+          name: cibw-wheels-${{ matrix.os }}
+          path: wheelhouse/*.whl
