+++ .github/workflows/ci.yaml
@@ -27,17 +27,13 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, windows-latest, macos-13, macos-14]
-        python-version: ['3.10','3.11','3.12','3.13']
+        os: [ubuntu-latest, windows-latest, macos-14] # macos-13,
   
     steps:
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
           submodules: true
-      - uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
       - uses: pypa/cibuildwheel@v2.21
       - name: Upload wheels
         uses: actions/upload-artifact@v4
