### 2025-01-07 bf7256b
try building only for 3.10
