+++ .github/workflows/ci.yaml
@@ -24,10 +24,12 @@ jobs:
   build_wheels:
     name: Wheel on ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
+    env: 
+      CIBW_BUILD: cp310-*
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, windows-latest, macos-14] # macos-13,
+        os: [ubuntu-latest, macos-14] # macos-13, macos-13, windows-latest
   
     steps:
       - uses: actions/checkout@v4
@@ -40,3 +42,5 @@ jobs:
         with:
           name: cibw-wheels-${{ matrix.os }}
           path: wheelhouse/*.whl
+      - name: Download all workflow run artifacts
+        uses: actions/download-artifact@v4
