+++ pyproject.toml
@@ -13,7 +13,7 @@ requires-python = ">3.10"
 dynamic = ["dependencies", "optional-dependencies"]
 
 [tool.setuptools]
-packages=['hetgpy','tests','docs','eigen']
+packages=['hetgpy','tests','docs']
 [tool.setuptools.dynamic]
 dependencies = {file = ["requirements.txt"]}
 optional-dependencies = {dev = { file = ["requirements_dev.txt"] }}
