### 2025-01-07 c509d60
restrict ci to auto64
