+++ .github/workflows/ci.yaml
@@ -25,7 +25,8 @@ jobs:
     name: Wheel on ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
     env: 
-      CIBW_BUILD: cp310-*
+      CIBW_BUILD: cp310-* cp311-* cp312-* cp13-*
+      CIBW_ARCHS: auto64
     strategy:
       fail-fast: false
       matrix:
