+++ pyproject.toml
@@ -13,7 +13,11 @@ requires-python = ">3.10"
 dynamic = ["dependencies", "optional-dependencies"]
 
 [tool.setuptools]
-packages=['hetgpy','tests','docs']
+packages=['hetgpy','docs']
+
+[tool.setuptools.package-data]
+"*" = ["*.yaml", "*.mat",".R",""]
+
 [tool.setuptools.dynamic]
 dependencies = {file = ["requirements.txt"]}
 optional-dependencies = {dev = { file = ["requirements_dev.txt"] }}
