+++ .github/workflows/ci.yaml
@@ -27,6 +27,8 @@ jobs:
     env: 
       CIBW_BUILD: cp310-* cp311-* cp312-* cp13-*
       CIBW_ARCHS: auto64
+      CIBW_TEST_REQUIRES: pytest
+      CIBW_TEST_COMMAND: "pytest hetgpy/tests/*.py"
     strategy:
       fail-fast: false
       matrix:
