### 2025-01-07 7a517e6
run pytest in CI
