### 2025-01-07 3650c80
run first directory of tests
