+++ .github/workflows/ci.yaml
@@ -28,7 +28,7 @@ jobs:
       CIBW_BUILD: cp310-* cp311-* cp312-* cp13-*
       CIBW_ARCHS: auto64
       CIBW_TEST_REQUIRES: pytest
-      CIBW_TEST_COMMAND: "pytest {project}/tests"
+      CIBW_TEST_COMMAND: "pytest {project}/tests/*.py"
     strategy:
       fail-fast: false
       matrix:
