### 2025-01-08 14caeec
remove pytest from cibuild
