+++ .github/workflows/ci.yaml
@@ -27,9 +27,6 @@ jobs:
     env: 
       CIBW_BUILD: cp310-* cp311-* cp312-* cp13-*
       CIBW_ARCHS: auto64
-      CIBW_TEST_REQUIRES: pytest
-      CIBW_TEST_COMMAND: "python -m pytest {project}/tests/test*.py"
-      CIBW_TEST_EXTRAS: ""
     strategy:
       fail-fast: false
       matrix:
