+++ .github/workflows/ci.yaml
@@ -138,6 +138,5 @@ jobs:
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
         repository-url: https://test.pypi.org/legacy/
-
-      - name: Download all workflow run artifacts
-        uses: actions/download-artifact@v4
+    - name: Download all workflow run artifacts
+      uses: actions/download-artifact@v4
