+++ .github/workflows/ci.yaml
@@ -32,7 +32,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-14] # macos-13, macos-13, windows-latest
+        os: [ubuntu-latest, macos-14, windows-latest] # macos-13, macos-13
   
     steps:
       - uses: actions/checkout@v4
