+++ .github/workflows/ci.yaml
@@ -69,6 +69,8 @@ jobs:
         merge-multiple: true
     - name: Publish distribution 📦 to PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
+      with: 
+        skip-existing: true
 
   github-release:
     name: >-
