+++ .github/workflows/run-tests-and-examples.yml
@@ -8,12 +8,12 @@ env:
 
 jobs:
   build_wheels:
-    name: Build wheels on ${{ matrix.os }}
+    name: Build wheels on ${{ matrix.os }} for $${ matrix.python-version }}
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
-        os: ['ubuntu-latest'] # [ubuntu-latest, windows-latest, macos-13, macos-14]
+        os: ['ubuntu-latest', 'macos-14','windows-latest'] # [ubuntu-latest, windows-latest, macos-13, macos-14]
         python-version: ['3.10', '3.11', '3.12','3.13']
 
     steps:
