+++ .github/workflows/run-tests-and-examples.yml
@@ -19,11 +19,6 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: actions/checkout@v4
-      - name: Setup R
-        uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: '4.3'
-      - run: Rscript -e 'print("hello")'
       # Used to host cibuildwheel
       - uses: actions/setup-python@v4
         with:
@@ -31,7 +26,7 @@ jobs:
       - name: Install dependencies
         run: python -m pip install -r requirements.txt
       - name: Install dev dependencies
-        run: python -m pip install -r requirements_dev.txt
+        run: python -m pip install -r requirements_testing.txt
       - name: Handle eigen
         run: git submodule update --init --recursive
       - name: Build package
