### 2025-01-08 0cd447f
only run basic integration tests
