### 2025-01-08 880c7fe
run basic tests
