+++ .github/workflows/run-basic-tests.yaml
@@ -1,3 +1,4 @@
+# run basic tests on push
 name: Build and Test
 
 on: push
@@ -13,8 +14,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ['windows-latest'] # [ubuntu-latest, windows-latest, macos-14]
-        python-version: ['3.10'] #, '3.11', '3.12','3.13']
+        os: ['ubuntu-latest'] # [ubuntu-latest, windows-latest, macos-14]
+        python-version: ['3.10','3.11', '3.12','3.13'] #, '3.11', '3.12','3.13']
 
     steps:
       - uses: actions/checkout@v4
@@ -33,11 +34,3 @@ jobs:
         run: pip install -e .
       - name: Test
         run: pytest tests/basic # run basic tests (not those in rpy2 folder)
-      - name: Run Intro Example
-        run: jupyter execute examples/00-hetGPy-Intro.ipynb
-      - name: Run Motorcycle Example
-        run: jupyter execute examples/01-mcycle-example.ipynb
-      #- name: Run SIR Example
-      #  run: jupyter execute examples/02-SIR-example.ipynb
-      - name: Download all workflow run artifacts
-        uses: actions/download-artifact@v4
