### 2025-01-08 49595ae
pin rpy2 version for testing
