### 2025-01-08 12c21e7
set R_HOME on windows
