### 2025-01-08 0f1834e
Create run-tests-all-platforms.yaml
