+++ .github/workflows/run-tests-all-platforms.yaml
@@ -1,0 +1,36 @@
+# run basic tests on push
+name: Run Basic Tests on All Platforms
+
+on: workflow_dispatch
+
+env:
+  OPENBLAS_NUM_THREADS: 1
+  OMP_NUM_THREADS: 1
+
+jobs:
+  build_wheels:
+    name: Build wheels on ${{ matrix.os }} for ${{ matrix.python-version }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ['ubuntu-latest','windows-latest','macos-14'] # [ubuntu-latest, windows-latest, macos-14]
+        python-version: ['3.10','3.11', '3.12','3.13'] #, '3.11', '3.12','3.13']
+
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/checkout@v4
+      # Used to host cibuildwheel
+      - uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: python -m pip install -r requirements.txt
+      - name: Install dev dependencies
+        run: python -m pip install -r requirements_testing.txt
+      - name: Handle eigen
+        run: git submodule update --init --recursive
+      - name: Build package
+        run: pip install -e .
+      - name: Test
+        run: pytest tests/basic # run basic tests (not those in rpy2 folder)
