### 2025-01-08 c03fd6d
run all tests on ubuntu only
