+++ .github/workflows/run-all-tests.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ['ubuntu-latest', 'windows-latest','macos-14'] # [ubuntu-latest, windows-latest, macos-13, macos-14]
+        os: ['ubuntu-latest'] # [ubuntu-latest, windows-latest, macos-13, macos-14]
         python-version: ['3.10','3.11','3.12','3.13']
 
     steps:
