### 2025-01-08 55c83c3
cleanup install script
