### 2025-01-08 davidogara/hetGPy#22 @davidogara
No worries @matthewfeickert, and I hope you are feeling better! I was able to address your two requests:

1. Added `ci.yaml` to the workflows, which builds wheels for the major platforms and supported python versions. I was able to still ship `Eigen` with the source dist, but remove it from the wheels (since you don't need it once the binaries are compiled-- which saved quite a bit of space, since Eigen is a somewhat large package)
2. In the README, I removed the discussion of "pure" and rephrased to say that the package aims to have few Python and external dependencies https://github.com/davidogara/hetGPy/blob/main/README.md?plain=1#L5-L12

I also ran the CI build and uploaded version 1.0.1 to pypi which you can see here: https://pypi.org/project/hetgpy/1.0.1/#files

Also, thanks so much for pointing to the `cibuildwheel` docs, I found those super helpful.

Please let me know any thoughts or feedback you have!
