### 2025-01-14 davidogara/hetGPy#22 @matthewfeickert
Sorry for the lack of response so far @davidogara &mdash; I've had some tight time constraints this week. I hope to review and respond to your nice work in the coming days.
