### 2025-01-14 davidogara/hetGPy#22 @davidogara
No problem @matthewfeickert! Thanks for letting me know. Not sure if this is relevant, but I did fix a small bug in https://github.com/davidogara/hetGPy/blob/main/hetgpy/src/matern.cpp

and uploaded a new patch version to pypi (https://pypi.org/project/hetgpy/1.0.2/)
