### 2025-01-27 davidogara/hetGPy#22 @matthewfeickert
@davidogara sorry for the very long delay, but this all is very nice work! You've nicely resolved this issue and I'm happy to close it and carry on with moving through the editor checks. 👍
