### 2025-01-27 davidogara/hetGPy#24 @matthewfeickert
[JOSS Review] Paper comments
This is in relation to [editor comments post review](https://github.com/openjournals/joss-reviews/issues/7518#issuecomment-2614994680) for https://github.com/openjournals/joss-reviews/issues/7518:

* "object orientated" should be "object oriented".

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.md?plain=1#L44

* In

> Its object orientated framework, high-level functionality, and third-party libraries such as numpy (Harris et al., 2020), scipy (Virtanen et al., 2020), PyTorch (Paszke et al., 2019) and scikit-learn (Pedregosa et al., 2011) make it a powerful tool for academic and industry professionals alike.

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.md?plain=1#L44

the usefulness of Python libraries is already well known, so I don't think that these need to be reintroduced here. It should be sufficient to just cite them as you mention them later on in the paper, as you do with NumPy, SciPy, and PyTorch (though please make sure to typeset the library names as they are given and not as you would import them). You don't make reference to scikit-learn elsewhere in the paper, so this can be dropped.

* From

> Our experiments indicate `hetGPy` is able to learn response surfaces efficiently, and in the case of high replication

onwards

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.md?plain=1#L51-L58

is this information critical to the "Statement of Need"? If not, I would just suggest having this information in your docs if you don't already, and removing it from here. I think the point has already been made before hand and as hetGPy doesn't make explicit statement about performance comparisons to PyTorch part of its main focus I think that these can be removed from the paper.

* Off the top of my head I don't know if the BibTeX system used for JOSS will pick up and render ISBN (**edit**: [JOSS does expect it](https://github.com/search?q=org%3Aopenjournals%20%22ISBN%22&type=code)), but please add and check that for books that you cite:

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.bib#L233-L238

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.bib#L186-L192

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.bib#L216-L224

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.bib#L226-L231

You can [check the renders using GitHub Actions or Docker](https://joss.readthedocs.io/en/latest/paper.html#checking-that-your-paper-compiles) without having to ask for a rerender from editorialbot.

* The rendering of

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.bib#L50-L55

will probably look better using `misc` instead of `article` and using `howpublished` instead of `url`.

* Some citations end up cutting off authors at some point, like for

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.bib#L462

and a few others. In these situations it is fine to use et al.

* The author name letter accents for

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.bib#L398

are not getting properly rendered here. Can you amend these from what the reference from the journal has, where they are for some reason using `\textbackslash`, which isn't needed?

---

Not relevant for publication, but throwing this in as well:

https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/docs/index.rst?plain=1#L31

has an extra backtick at the end which shows up in the render.
