### 2025-01-27 139d3cf
remove Zotero file paths
