+++ paper/paper.bib
@@ -12,7 +12,6 @@
 	year = {2021},
 	note = {Publisher: Public Library of Science},
 	pages = {e1009149},
-	file = {Full Text:/Users/davidogara/Zotero/storage/R6DFYUGV/Kerr et al. - 2021 - Covasim An agent-based model of COVID-19 dynamics.pdf:application/pdf},
 }
 
 @article{binois_hetgp_2021,
@@ -28,7 +27,6 @@
 	journal = {Journal of Statistical Software},
 	author = {Binois, Mickaël and Gramacy, Robert B.},
 	year = {2021},
-	file = {Binois and Gramacy - 2021 - hetGP  Heteroskedastic Gaussian Process Mo.pdf:/Users/davidogara/Zotero/storage/257WNMGS/Binois and Gramacy - 2021 - hetGP  Heteroskedastic Gaussian Process Mo.pdf:application/pdf},
 }
 
 @article{ogara_traceomicron_2023,
@@ -44,7 +42,6 @@
 	month = apr,
 	year = {2023},
 	pages = {2300147},
-	file = {O'Gara et al_2023_TRACE‐Omicron.pdf:/Users/davidogara/Library/CloudStorage/Box-Box/Zotero/O'Gara et al_2023_TRACE‐Omicron.pdf:application/pdf},
 }
 
 @article{Gautier:2008,
@@ -84,7 +81,6 @@
 	month = jul,
 	year = {2021},
 	pages = {210506},
-	file = {Aylett-Bullock et al_2021_J span style=font-variant.pdf:/Users/davidogara/Zotero/storage/MHLXPBPS/Aylett-Bullock et al_2021_J span style=font-variant.pdf:application/pdf},
 }
 
 @article{lazaridis_genetic_2022,
@@ -102,8 +98,7 @@
 	year = {2022},
 	note = {Publisher: American Association for the Advancement of Science},
 	pages = {eabm4247},
-	file = {Full Text PDF:/Users/davidogara/Zotero/storage/JPCMBTVM/Lazaridis et al. - 2022 - The genetic history of the Southern Arc A bridge .pdf:application/pdf},
-}
+	}
 
 @article{binois_practical_2018,
 	title = {Practical {Heteroscedastic} {Gaussian} {Process} {Modeling} for {Large} {Simulation} {Experiments}},
@@ -120,7 +115,6 @@
 	note = {Publisher: Taylor \& Francis},
 	keywords = {Input-dependent noise, Replication, Stochastic kriging, Woodbury formula},
 	pages = {808--821},
-	file = {Submitted Version:/Users/davidogara/Zotero/storage/JQM3XXBA/Binois et al. - 2018 - Practical Heteroscedastic Gaussian Process Modelin.pdf:application/pdf},
 }
 
 @article{binois_replication_2019,
@@ -139,7 +133,6 @@
 	note = {Publisher: Taylor \& Francis},
 	keywords = {Input-dependent noise, Computer experiment, Gaussian process, Lookahead, Replicated observations, Surrogate model},
 	pages = {7--23},
-	file = {Submitted Version:/Users/davidogara/Zotero/storage/EX35CVWC/Binois et al. - 2019 - Replication or Exploration Sequential Design for .pdf:application/pdf},
 }
 
 @misc{baker_analyzing_2020,
@@ -156,7 +149,6 @@
 	note = {arXiv:2002.01321 [stat]},
 	keywords = {Statistics - Methodology},
 	annote = {Comment: 48 pages, 8 figures},
[…]
