### 2025-01-27 c5177ba
make case study section
