+++ paper/paper.bib
@@ -178,6 +178,7 @@ A rapid vaccination rollout can allow the sooner relaxation of NPIs, however ong
 	title = {Bayesian {Optimization}},
 	publisher = {Cambridge University Press},
 	author = {Garnett, Roman},
+	isbn = {9781108425780},
 	year = {2023},
 }
 
@@ -208,6 +209,7 @@ Publisher: Nature Publishing Group},
 	publisher = {Chapman Hall/CRC},
 	author = {Gramacy, Robert B.},
 	year = {2020},
+	isbn = {9780367415426},
 	annote = {http://bobby.gramacy.com/surrogates/},
 }
 
@@ -215,6 +217,7 @@ Publisher: Nature Publishing Group},
 	title = {Modeling and {Simulation} in {Python}},
 	publisher = {CRC Press},
 	author = {Kinser, Jason M},
+	isbn = {978-1032116488},
 	year = {2022},
 }
 
@@ -222,6 +225,7 @@ Publisher: Nature Publishing Group},
 	title = {Modeling and {Simulation} in {Python}: {An} {Introduction} for {Scientists} and {Engineers}},
 	publisher = {No Starch Press},
 	author = {Downey, Allen B},
+	introduction = {9781718502161},
 	year = {2023},
 }
 
