+++ paper/paper.bib
@@ -225,7 +225,7 @@ Publisher: Nature Publishing Group},
 	title = {Modeling and {Simulation} in {Python}: {An} {Introduction} for {Scientists} and {Engineers}},
 	publisher = {No Starch Press},
 	author = {Downey, Allen B},
-	introduction = {9781718502161},
+	isbn = {9781718502161},
 	year = {2023},
 }
 
@@ -316,7 +316,7 @@ Publisher: Nature Publishing Group},
 @book{forrester_engineering_2008,
 	edition = {1},
 	title = {Engineering {Design} via {Surrogate} {Modelling}: {A} {Practical} {Guide}},
-	isbn = {978-0-470-06068-1 978-0-470-77080-1},
+	isbn = {978-0-470-06068-1},
 	shorttitle = {Engineering {Design} via {Surrogate} {Modelling}},
 	url = {https://onlinelibrary.wiley.com/doi/book/10.1002/9780470770801},
 	language = {en},
@@ -465,7 +465,7 @@ Publisher: Nature Publishing Group},
 }
 
 
-@book{ushey_reticulate_2023,
+@misc{ushey_reticulate_2023,
 	annote = {R package version 1.30},
 	author = {Ushey, Kevin and Allaire, J. J. and Tang, Yuan},
 	title = {reticulate: {Interface} to '{Python}'},
@@ -499,6 +499,7 @@ Publisher: Nature Publishing Group},
 	author = {Harville, David A.},
 	publisher = {Springer},
 	title = {Matrix {Algebra} {From} a {Statistician}'s {Perspective}},
+	isbn = {978-0-387-22677-4},
 	year = {1997}}
 
 @software{jax2018github,
