+++ paper/paper.bib
@@ -44,10 +44,11 @@
 	pages = {2300147},
 }
 
-@article{Gautier:2008,
+@misc{Gautier:2008,
   title={rpy2: A Simple and Efficient Access to R from Python},
   author={Gautier, L},
-  journal={URL http://rpy. sourceforge. net/rpy2. html},
+  howpublished = {Web page},
+  url = {http://rpy. sourceforge. net/rpy2. html},
   year={2008}
 }
 @article{aylett-bullock_j_2021,
@@ -382,7 +383,7 @@ Publisher: Nature Publishing Group},
 	number = {40},
 	urldate = {2024-08-01},
 	journal = {Journal of Machine Learning Research},
-	author = {Matthews, Alexander G. de G. and Wilk, Mark van der and Nickson, Tom and Fujii, Keisuke and Boukouvalas, Alexis and Le\{{\textbackslash}'o\}n-Villagr\{{\textbackslash}'a\}, Pablo and Ghahramani, Zoubin and Hensman, James},
+	author = {Matthews, Alexander G. de G. and Wilk, Mark van der and Nickson, Tom and Fujii, Keisuke and Boukouvalas, Alexis and Le{\'o}n-Villagr{\'a}, Pablo and Ghahramani, Zoubin and Hensman, James},
 	year = {2017},
 	pages = {1--6},
 }
