### 2025-01-28 4d26a9e
update paper and bib files
