+++ paper/paper.md
@@ -49,7 +49,7 @@ The state of the art for GPR in Python is `GPyTorch` [@gardner_gpytorch_2018], f
  
 # Additional Features and Case Study
 
-hetGPy also has two intermediate goals: (1) efficient computations, accomplished via implementation on numpy arrays and (2) minimal dependencies, the core of which are `NumPy` for efficient array-based computation and `SciPy` which contains the definitive implementation of the L-BFGS-B algorithm in Fortran [@morales_remark_2011;@byrd_limited_1995] used for maximum likelihood estimation of hyperparameters. Our experiments indicate `hetGPy` is able to learn response surfaces efficiently, and in the case of high replication, do so on CPUs more efficiently than the default implementation in `GPyTorch`, as shown for a suite of test problems in Table 1. The test problems contain 1,000 unique design locations and an average of 5 replications per location, meaning $n=1,000$ and $N \approx 5,000$ which shows why `hetGPy` is far more efficient in this setting.  As a comparator, we also conduct a set of experiments using stochastic kriging (SK) [@ankenman_stochastic_2010], a precursor method to @binois_practical_2018 that also allows for maximum likelihood estimation with replication, and under the case of homoskedasticity, is nearly equivalent to homoscedastic GPR in `hetGP` and `hetGPy` [@gramacy_surrogates_2020]. We implement SK with a custom `GPyTorch` likelihood that accounts for replication under homoskedasticity. Specifically, given a dataset X with unique designs $(X_1,...,X_k)$ each replicated $(n_1,...,n_k)$ times, we pre-average the outputs $Y_i$ at each unique input location, and then estimate the diagonal of the noise matrix as  $(\sigma^2⁄n_i )$. We see that for the SK case, `hetGPy` and `GPyTorch` have training times on a similar order of magnitude. The package `hetGPy` is under active development and is well-posed to engage with the wider Python community for future extension such as arbitrary kernel functions with auto-differentiation methods facilitated by `PyTorch` or `jax` [@jax2018github].
[…]
