+++ paper/paper.bib
@@ -7,7 +7,8 @@
 	abstract = {Author summary Mathematical models have played an important role in helping countries around the world decide how to best tackle the COVID-19 pandemic. In this paper, we describe a COVID-19 model, called Covasim (COVID-19 Agent-based Simulator), that we developed to help answer these questions. Covasim can be tailored to the local context by using detailed data on the population (such as the population age distribution and number of contacts between people) and the epidemic (such as diagnosed cases and reported deaths). While Covasim can be used to explore theoretical research questions or to make projections, its main purpose is to evaluate the effect of different interventions on the epidemic. These interventions include physical interventions (mobility restrictions and masks), diagnostic interventions (testing, contact tracing, and quarantine), and pharmaceutical interventions (vaccination). Covasim is open-source, written in Python, and comes with extensive documentation, tutorials, and a webapp to ensure it can be used as easily and broadly as possible. In partnership with local stakeholders, Covasim has been used to answer policy and research questions in more than a dozen countries, including India, the United States, Vietnam, and Australia.},
 	number = {7},
 	journal = {PLOS Computational Biology},
-	author = {Kerr, Cliff C. and Stuart, Robyn M. and Mistry, Dina and Abeysuriya, Romesh G. and Rosenfeld, Katherine and Hart, Gregory R. and Núñez, Rafael C. and Cohen, Jamie A. and Selvaraj, Prashanth and Hagedorn, Brittany and George, Lauren and Jastrzębski, Michał and Izzo, Amanda S. and Fowler, Greer and Palmer, Anna and Delport, Dominic and Scott, Nick and Kelly, Sherrie L. and Bennette, Caroline S. and Wagner, Bradley G. and Chang, Stewart T. and Oron, Assaf P. and Wenger, Edward A. and Panovska-Griffiths, Jasmina and Famulare, Michael and Klein, Daniel J.},
+	%author = {Kerr, Cliff C. and Stuart, Robyn M. and Mistry, Dina and Abeysuriya, Romesh G. and Rosenfeld, Katherine and Hart, Gregory R. and Núñez, Rafael C. and Cohen, Jamie A. and Selvaraj, Prashanth and Hagedorn, Brittany and George, Lauren and Jastrzębski, Michał and Izzo, Amanda S. and Fowler, Greer and Palmer, Anna and Delport, Dominic and Scott, Nick and Kelly, Sherrie L. and Bennette, Caroline S. and Wagner, Bradley G. and Chang, Stewart T. and Oron, Assaf P. and Wenger, Edward A. and Panovska-Griffiths, Jasmina and Famulare, Michael and Klein, Daniel J.},
+	author = {Kerr, Cliff C. et al},
 	month = jul,
 	year = {2021},
 	note = {Publisher: Public Library of Science},
@@ -94,7 +95,8 @@
 	number = {6609},
 	urldate = {2024-02-13},
 	journal = {Science},
[…]
