### 2025-01-28 davidogara/hetGPy#24 @davidogara
Thanks so much for sending these @matthewfeickert. I was able to address almost of all of them without issue, and organized them into a pull request here, if you'd like to take a look: https://github.com/davidogara/hetGPy/pull/25

If it looks good to you I will merge to main.

The only follow I have is from your question on:
https://github.com/davidogara/hetGPy/blob/1a68c2f39108e2d3a36a5cfc3b03c9ccfb6f6d5b/paper/paper.md?plain=1#L51-L58

My thought was to start a new section header called "Additional Features and Case Study" since we talk a bit about the additional goals for `hetGPy` and do a simple experiment comparing to `GPyTorch`. The reason I'd be inclined to include this section is that we wanted to compare the performance to at least one widely-used library (to help users understand when `hetGPy` might be well-suited for a particular task), and this was the section of the paper that our two reviewers had the most feedback on.

A different option would be to instead call that section a "Vignette" which I saw done in the `sciris` JOSS paper, if you think that would be more accurate to what we did (https://joss.theoj.org/papers/10.21105/joss.05076).

Would definitely appreciate any thoughts you have. Thanks again for the paper edits, and for the tips and tricks for fixing the bib file/compiling the pdf.
