### 2025-01-31 davidogara/hetGPy#24 @matthewfeickert
> My thought was to start a new section header called "Additional Features and Case Study" since we talk a bit about the additional goals for `hetGPy` and do a simple experiment comparing to `GPyTorch`. The reason I'd be inclined to include this section is that we wanted to compare the performance to at least one widely-used library (to help users understand when `hetGPy` might be well-suited for a particular task), and this was the section of the paper that our two reviewers had the most feedback on.

That sounds fine to me. 👍 Thanks for theexplanationn of why you wanted to show this.
