### 2025-01-31 3002796
Update paper/paper.bib
Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
