+++ paper/paper.bib
@@ -48,8 +48,7 @@
 @misc{Gautier:2008,
   title={rpy2: A Simple and Efficient Access to R from Python},
   author={Gautier, L},
-  howpublished = {Web page},
-  url = {http://rpy. sourceforge. net/rpy2. html},
+  url = {https://github.com/rpy2/rpy2},
   year={2008}
 }
 @article{aylett-bullock_j_2021,
