### 2025-01-31 davidogara/hetGPy#24 @davidogara
Thank you! I'll go ahead and close the issue.
