### 2025-01-31 #7518 @matthewfeickert
Nice @davidogara. Once you have an archive with a DOI for [v1.0.3](https://github.com/davidogara/hetGPy/releases/tag/v1.0.3) (https://github.com/openjournals/joss-reviews/issues/7518#issuecomment-2615076353) and have verified things on your side (https://github.com/openjournals/joss-reviews/issues/7518#issuecomment-2614994680) we can recommend publication. 👍
