+++ paper/paper.bib
@@ -220,6 +220,7 @@ Publisher: Nature Publishing Group},
 	publisher = {CRC Press},
 	author = {Kinser, Jason M},
 	isbn = {978-1032116488},
+	doi = {10.1201/9781003226581},
 	year = {2022},
 }
 
