### 2025-01-31 #7518 @davidogara
@matthewfeickert everything in the checklist looks good to me!

The Zenodo archive is here: https://doi.org/10.5281/zenodo.14782113
