### 2025-01-31 #7518 @matthewfeickert
> The Zenodo archive is here: https://doi.org/10.5281/zenodo.14782113

@davidogara Ah, the archive is supposed to be a snapshot of the repository at the tag used for the paper, not release build products. (c.f. https://doi.org/10.5281/zenodo.1169739 as an example)

Can you upload the repository state at https://github.com/davidogara/hetGPy/tree/v1.0.3 to the existing archive, and then open a ticket with Zenodo to delete https://doi.org/10.5281/zenodo.14782113 ?
