### 2025-01-31 #7518 @matthewfeickert
> I re-uploaded only the source code -- the updated doi is https://doi.org/10.5281/zenodo.14782341

@davidogara Nice, thanks.

The archive should have the same license as the repository. https://doi.org/10.5281/zenodo.14782341 currently has

> Creative Commons Attribution 4.0 International

listed, but hetGPy is [GNU Lesser General Public License v2.1](https://github.com/davidogara/hetGPy/blob/6b0b352211cb9f1c8f80fddb93e152f2b93c0da3/LICENSE). The metadata for the archives is editable (when logged in with your Zenodo account there will be an edit button on the right hand side of the screen for https://zenodo.org/records/14782341), so can you update the license listed for https://doi.org/10.5281/zenodo.14782341?
