### 2025-01-31 #7518 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
