### 2025-01-31 #7518 @matthewfeickert
@davidogara to avoid confusion if you can also open up a ticket with Zenodo (https://zenodo.org/support) and request that _just_ https://doi.org/10.5281/zenodo.14782113 be deleted that would be great. 👍
