### 2025-01-31 #7518 @davidogara
@matthewfeickert just following up-- Zenodo did delete https://doi.org/10.5281/zenodo.14782113 without issue
