### 2025-02-04 #7518 @matthewfeickert
> I checked the proof and the XML files and everything looked good to me. Were there any other items you needed us to address before any next steps?

Nope. I've already added the https://github.com/openjournals/joss-reviews/labels/recommend-accept label, so we're now just waiting for the track Editor-in-Chief to review, approve, and publish.
