### 2025-02-04 #7518 @davidogara
> > Were there any other items you needed us to address before any next steps?
>
> Nope. I've already added the [ recommend-accept](https://github.com/openjournals/joss-reviews/labels/recommend-accept) Papers recommended for acceptance in JOSS. label, so we're now just waiting for the [ Track: 5 (DSAIS)](https://github.com/openjournals/joss-reviews/labels/Track%3A%205%20%28DSAIS%29) Data Science, Artificial Intelligence, and Machine Learning track Editor-in-Chief to review, approve, and publish.

Sounds great, thank you for letting me know!
