+++ paper/paper.bib
@@ -46,7 +46,7 @@
 }
 
 @misc{Gautier:2008,
-  title={rpy2: A Simple and Efficient Access to R from Python},
+  title={rpy2: A Simple and Efficient Access to {R} from {P}ython},
   author={Gautier, L},
   url = {https://github.com/rpy2/rpy2},
   year={2008}
