### 2025-02-05 #7518 @davidogara
Thanks @crvernon! I just made that change in https://github.com/davidogara/hetGPy/commit/c686bcb502d605722f3ebdd1ce54d8dc00276cc8
