### 2025-02-05 #7518 @crvernon
@davidogara could you please remove the backtick formatting of your package name in your paper? It is causing a formatting issue with how JOSS displays your title in our published papers list.
