### 2025-02-05 1e2c025
Create CITATION.cff
