+++ CITATION.cff
@@ -1,0 +1,48 @@
+cff-version: "1.2.0"
+authors:
+- family-names: O'Gara
+  given-names: David
+  orcid: "https://orcid.org/0000-0002-1957-400X"
+- family-names: Binois
+  given-names: Mickaël
+  orcid: "https://orcid.org/0000-0002-7225-1680"
+- family-names: Garnett
+  given-names: Roman
+  orcid: "https://orcid.org/0000-0002-0152-5453"
+- family-names: Hammond
+  given-names: Ross A
+  orcid: "https://orcid.org/0009-0005-1046-9296"
+contact:
+- family-names: O'Gara
+  given-names: David
+  orcid: "https://orcid.org/0000-0002-1957-400X"
+doi: 10.5281/zenodo.14782341
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: O'Gara
+    given-names: David
+    orcid: "https://orcid.org/0000-0002-1957-400X"
+  - family-names: Binois
+    given-names: Mickaël
+    orcid: "https://orcid.org/0000-0002-7225-1680"
+  - family-names: Garnett
+    given-names: Roman
+    orcid: "https://orcid.org/0000-0002-0152-5453"
+  - family-names: Hammond
+    given-names: Ross A
+    orcid: "https://orcid.org/0009-0005-1046-9296"
+  date-published: 2025-02-05
+  doi: 10.21105/joss.07518
+  issn: 2475-9066
+  issue: 106
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 7518
+  title: "hetGPy: Heteroskedastic Gaussian Process Modeling in Python"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.07518"
+  volume: 10
+title: "`hetGPy`: Heteroskedastic Gaussian Process Modeling in Python"
