### 2025-02-05 #7518 @davidogara
@crvernon yep! just made that change in [f3c4cca](https://github.com/davidogara/hetGPy/commit/f3c4ccaa9cb935ed1a0458b43c019f422b9de1ae)
